% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_axis_labels}
\alias{add_axis_labels}
\title{Add Axis Labels to a Plot}
\usage{
add_axis_labels(gg_obj, xlab = NULL, ylab = NULL)
}
\arguments{
\item{gg_obj}{A \code{ggplot} object}

\item{xlab}{The x axis label}

\item{ylab}{The y axis label}
}
\value{
The modified \code{ggplot} object
}
\description{
\code{add_axis_labels()} adds axis labels to a \code{ggplot} object.
}
\keyword{internal}
