% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_covid_events}
\alias{add_covid_events}
\title{Add a Set of Events From the Shelby County COVID Response to a Plot}
\usage{
add_covid_events(gg_obj, lab_y, ...)
}
\arguments{
\item{gg_obj}{A \code{ggplot} object}

\item{lab_y}{The y coordinate of the event label}

\item{...}{Additional arguments to pass to
\code{\link[coviData:add_event]{add_event()}}}
}
\value{
The modified \code{ggplot} object
}
\description{
\code{add_covid_events()} adds a set of events from the Shelby County COVID-19
response to a \code{ggplot} object.
}
\section{Current Events}{


\itemize{
\item{2020-03-24: Safer-at-Home}
\item{2020-05-04: Phase 1 Reopen}
\item{2020-05-18: Phase 2 Reopen}
\item{2020-07-08: Bars Close, Mask Directive}
\item{2020-09-22: Bars Reopen}
\item{2020-09-29: End State Restrictions}
\item{2020-11-23: Face Mask Order 3, Dining Restrictions}
\item{2020-12-26: Face Mask Order 4, Safer-at-Home}
\item{2021-01-23: Safer-at-Home Lifted}
\item{2021-02-20: Broaden Safety Measures}
\item{2021-03-20: Lift More Restrictions}
\item{2021-04-17: Emphasize Vaccination}
}
}

\keyword{internal}
