% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_event}
\alias{add_event}
\title{Add an Event Marker on a Plot}
\usage{
add_event(
  gg_obj,
  date,
  lab,
  lab_y,
  hjust = 0,
  vjust = 0,
  angle = -90,
  face = c("bold", "italic", "bold.italic", "plain"),
  line = c("dashed", "solid", "blank", "dotted"),
  color = "grey30",
  ...
)
}
\arguments{
\item{gg_obj}{A \code{ggplot} object}

\item{date}{The date of the event}

\item{lab}{The event label}

\item{lab_y}{The y coordinate of the event label}

\item{hjust}{The horizontal alignment of the event label; defaults to \code{0}}

\item{vjust}{The vertical alignment of the event label; defaults to \code{0}}

\item{angle}{The angle of the event label; defaults to \code{-90}, which reads
"sideways" from top to bottom. This is a tradeoff that creates space on
most time series plots.}

\item{face}{The font face of the label text}

\item{line}{The type of vertical line to use}

\item{color}{Line and text color}

\item{...}{Additional arguments to pass to
\code{\link[ggplot2:geom_text]{geom_text()}}}
}
\value{
The modified \code{ggplot} object
}
\description{
\code{add_event()} adds a vertical line and a label denoting an event to
a \code{ggplot} object.
}
\keyword{internal}
