% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_scale_month}
\alias{add_scale_month}
\title{Add a Monthly X-Axis Scale to a Time-Based Plot}
\usage{
add_scale_month(gg_obj)
}
\arguments{
\item{gg_obj}{A \code{ggplot} object}
}
\value{
The modified \code{ggplot} object
}
\description{
\code{add_scale_month()} adds a monthly scale to a \code{ggplot} object when the
x axis is a date or datetime.
}
\keyword{internal}
