% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_title_caption}
\alias{add_title_caption}
\title{Add a Title, Subtitle, and/or Caption to a Plot}
\usage{
add_title_caption(gg_obj, title = NULL, subtitle = NULL, caption = NULL)
}
\arguments{
\item{gg_obj}{A \code{ggplot} object}

\item{title}{The title of the plot}

\item{subtitle}{The subtitle of the plot}

\item{caption}{The caption of the plot}
}
\value{
The modified \code{ggplot} object
}
\description{
\code{add_title_caption()} adds a title, subtitle, and/or caption with
standardized settings to a \code{ggplot} object.
}
\keyword{internal}
