% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_today}
\alias{add_today}
\title{Add a Marker for Today's Date to a Plot}
\usage{
add_today(
  gg_obj,
  lab_y,
  hjust = 0,
  vjust = 0,
  angle = 0,
  face = c("plain", "bold", "italic", "bold.italic"),
  line = c("dotted", "dashed", "solid", "blank")
)
}
\arguments{
\item{gg_obj}{A \code{ggplot} object}

\item{lab_y}{The y coordinate of the event label}

\item{hjust}{The horizontal alignment of the event label; defaults to \code{0}}

\item{vjust}{The vertical alignment of the event label; defaults to \code{0}}

\item{angle}{The angle of the event label; defaults to \code{0}}

\item{face}{The font face of the label text}

\item{line}{The type of vertical line to use}
}
\value{
The modified \code{ggplot} object
}
\description{
\code{add_today()} adds a dotted line and a "Today" label to a \code{ggplot} object.
}
\keyword{internal}
