% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_date.R
\name{archive_report_date}
\alias{archive_report_date}
\title{Archive \code{fst} Files of NBS Report Dates Based on Date of Creation}
\usage{
archive_report_date(
  data_dir = path_create("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/",
    "Sandbox data pull Final/"),
  archive_dir = path_create("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/",
    "Report Date Linelist/archive/")
)
}
\arguments{
\item{data_dir}{The directory containing the NBS data files to archive}

\item{archive_dir}{The location of the archive directory}
}
\value{
A \code{tibble} containing the data files paths, creation (birth) date,
and archive file paths for newly created archive data
}
\description{
\code{archive_report_date()} infers the date reported for NBS IDs using the date
of creation for the data files in \code{data_dir}. It saves the inferred report
dates in \code{archive_dir}.
}
