% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_investigations.R
\name{asg_join_nights_weekends}
\alias{asg_join_nights_weekends}
\title{Join Nights and Weekends Schedules to Teams Data}
\usage{
asg_join_nights_weekends(
  .data,
  path = path_create("V:/Administration/Schedules/Night&Weekend staff", ext = "xlsx"),
  sheet = "schedules"
)
}
\arguments{
\item{.data}{The output of \code{as_join_schedules()}}

\item{path}{The location of the nights and weekends scheduling workbook}

\item{sheet}{The sheet from which to read the schedule; either the sheet name
or the sheet number}
}
\value{
A \code{tibble} with a \code{schedule_custom} column added
}
\description{
\code{asg_join_nights_weekends()} loads individual schedules for investigators on
night or weekend shifts and joins them with the output of
\code{asg_join_schedules()}.
}
\seealso{
Other Case Assignment: 
\code{\link{asg_calc_schedules}()},
\code{\link{asg_join_schedules}()},
\code{\link{asg_load_nights_weekends}()},
\code{\link{asg_load_teams}()},
\code{\link{asg_parse_names}()},
\code{\link{asg_parse_nights_weekends}()},
\code{\link{asg_schedule_predefined}()},
\code{\link{asg_schedule}()}
}
\concept{Case Assignment}
