% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_bool}
\alias{assert_bool}
\title{Assert that an Object is Boolean}
\usage{
assert_bool(x, arg = NULL)
}
\arguments{
\item{x}{An object to test}

\item{arg}{Optional argument name to display in error messages. If not
provided, the input \code{x} is parsed to a name using \code{expr_label(enexpr(x))}.}
}
\value{
\code{x}, invisibly
}
\description{
\code{assert_bool()} checks whether an object is (scalar) Boolean using
\code{\link[rlang:is_bool]{is_bool()}}.
}
