% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_cols}
\alias{assert_cols}
\title{Assert That a Data Frame Selection Matches the Specified Type and Length}
\usage{
assert_cols(.data, ..., ptype = NULL, n = NULL)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g. a \code{tibble})}

\item{...}{\verb{<tidy-select>} One or more selection specifications for a set of
columns in the dataframe}

\item{ptype}{The prototype that the selected columns should match}

\item{n}{The number of columns that should be selected}
}
\value{
The input object, \code{.data}
}
\description{
\code{assert_cols()} checks whether columns selected using \code{...} match the given
prototype and/or number of columns. If so, it returns the data frame;
otherwise, it throws an error.
}
\keyword{internal}
