% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_dataframe}
\alias{assert_dataframe}
\title{Assert that an Object Is a Data Frame}
\usage{
assert_dataframe(x, arg = NULL)
}
\arguments{
\item{x}{An object to check}

\item{arg}{The name of the object; will be set to the name of the input if
left \code{NULL}. This is useful for creating more informative errors.}
}
\value{
The input, \code{x}
}
\description{
\code{assert_dataframe()} checks whether the input object is a data frame. If so,
it returns the object; if not, it throws an error.
}
\keyword{internal}
