% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-assertthat.R
\name{assertthat-helpers}
\alias{assertthat-helpers}
\title{User-Friendly Assertions for R}
\description{
The assertthat package provides helpers for making assertions about a object.
It is useful for checking assumptions and throwing errors if they are not
met. See \url{https://github.com/hadley/assertthat} for more information.
}
\keyword{internal}
