% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_ael}
\alias{check_ael}
\title{Check Two AEL Data Files for New or Removed Results}
\usage{
check_ael(
  .data = NULL,
  date1 = Sys.Date(),
  date2 = date1 - 1,
  directory = "V:/EPI DATA ANALYTICS TEAM/AEL Data/",
  date1_file = NULL,
  date2_file = NULL,
  overwrite_names_1 = FALSE,
  overwrite_names_2 = FALSE,
  encoding = TRUE,
  rtn_table = TRUE
)
}
\arguments{
\item{.data}{If data has already been read into R, supply here (not used yet)}

\item{date1}{The \code{Date} corresponding to the first file}

\item{date2}{The \code{Date} corresponding to the second file}

\item{directory}{The directory to look for AEL data in}

\item{date1_file}{Optionally, the name of the first file; useful if searching
for a file with non-standard naming}

\item{date2_file}{Optionally, the name of the second file; useful if
searching for a file with non-standard naming}

\item{overwrite_names_1}{Should the patient names in the first file be
pre-processed with \code{\link{process_names}}?}

\item{overwrite_names_2}{Should the patient names in the second file be
pre-processed with \code{\link{process_names}}?}

\item{encoding}{Should character encoding be converted to \code{UFT-8}?}

\item{rtn_table}{Should a \code{\link[gt]{gt}} table be returned for the
first file?}
}
\description{
\code{check_ael} reads two AEL data files and checks for results in each that are
not in the other. This is primarily useful if the files are sequential.
}
