% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_generic.R
\name{check_date_updated}
\alias{check_date_updated}
\title{Check Update Date in Data for Regions REDcap Project}
\usage{
check_date_updated(
  date = Sys.Date(),
  api_token = Sys.getenv("redcap_DFR_token"),
  region = "MSR",
  date_updated = "date_updated",
  quiet = FALSE,
  vac = FALSE
)
}
\arguments{
\item{date}{The date to check against; defaults to \code{Sys.Date()}, which is
probably what you want}

\item{api_token}{The API token for accessing the Data for Regions REDcap
project. This should be stored in an \emph{.Renviron} file; see
\link{env_vars} for details.}

\item{region}{The region to check for; exists to aid internal implementation
and should probably not be changed from default}

\item{date_updated}{The name of the record containing the date updated in
REDcap; this exists to aid internal implementation and should probably not
be changed from default}

\item{quiet}{By default, warnings are issued if date does not match, and a
message is issued if it does; should these be silenced?}

\item{vac}{Should vaccination data be checked (\code{TRUE}), or NBS data
(\code{FALSE}, the default)?}
}
\value{
\code{TRUE} if dates match, \code{FALSE} otherwise
}
\description{
\code{check_date_updated()} checks whether the supplied \code{date}
(today, by default) matches the \code{date_updated} record for the specified
region in the Data for Regions REDcap project.
}
\seealso{
\code{\link{data-for-regions-snapshots}}
}
