% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_deaths}
\alias{check_deaths}
\title{Compare Linelists from NBS and Surveillance Team and Report Differences}
\usage{
check_deaths(
  surveillance_file = "Working Copy Death  Epi.xlsx",
  nbs_file = "NBSDeathLineList.xlsx",
  directory = "V:/EPI DATA ANALYTICS TEAM/MORTALITY DATA/",
  surveillance_file_id = NULL,
  nbs_file_id = PATIENT_LOCAL_ID,
  save = FALSE
)
}
\arguments{
\item{surveillance_file}{The name of the file containing the surveillance
team linelist}

\item{nbs_file}{The name of the file containing the NBS exported linelist}

\item{directory}{The path to the directory containing both linelists}

\item{surveillance_file_id}{The NBS ID column in the surveillance file. If
\code{NULL}, \code{check_deaths()} assumes that this is the first column in the file.}

\item{nbs_file_id}{The NBS ID column in the NBS linelist.}

\item{save}{Should the results be saved to a file?}
}
\description{
Compare Linelists from NBS and Surveillance Team and Report Differences
}
