% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_linelist_dates}
\alias{check_linelist_dates}
\title{Check and Clean Dates in a Linelist}
\usage{
check_linelist_dates(
  .data,
  .collection_date = "collection_date",
  .report_date = "report_date",
  today = Sys.Date(),
  quiet = FALSE
)
}
\arguments{
\item{.data}{A linelist with one case/test per row}

\item{.collection_date}{A \code{Date} column corresponding to the date of sample
collection}

\item{.report_date}{A \code{Date} column corresponding to the date the sample
was reported}

\item{today}{The \code{Date} to consider as "today"}

\item{quiet}{Should messages be shown for each checking and removal step?}
}
\value{
The checked and cleaned linelist
}
\description{
\code{check_linelist_dates()} checks a linelist with collection and report dates
for errors and logical inconsistencies. It removes observations that fail
these tests.
}
