% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_date.R
\name{chr_to_dttm}
\alias{chr_to_dttm}
\title{Parse Dates in Character Format to Datetime Format}
\usage{
chr_to_dttm(
  x,
  tz = "UTC",
  orders = c("Omdy", "dOmy", "yOmd", "OmdyT", "dOmyT", "yOmdT", "TOmdy", "TdOmy",
    "TyOmd", "OmdyR", "dOmyR", "yOmdR", "Omdyr", "dOmyr", "yOmdr", "OmdyTz", "dOmyTz",
    "yOmdTz", "TOmdyz", "TdOmyz", "TyOmdz", "OmdyRz", "dOmyRz", "yOmdRz", "Omdyrz",
    "dOmyrz", "yOmdrz")
)
}
\arguments{
\item{x}{A vector of character dates}

\item{tz}{Optional timezone}

\item{orders}{The orders to use when parsing character vector with
\code{\link[lubridate:parse_date_time]{parse_date_time()}}}
}
\value{
A \code{POSIXct} vector
}
\description{
\code{chr_to_dttm} standardizes a datetime vector in character format and returns
a vector in \code{POSIXct} format.
}
