% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{clean_generic}
\alias{clean_generic}
\title{Standardize Variables, Handle Dates, and Remove Bad Columns}
\usage{
clean_generic(.data, string_to_factor = FALSE)
}
\description{
\code{clean_generic} standardizes column names, removes empty or constant
columns, converts variables to the appropriate data types (including oddly
formatted dates/datetimes), removes empty or constant columns, and optionally converts
strings to factors.
}
