% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coalesce_dupes}
\alias{coalesce_dupes}
\title{Coalesce Information in Duplicate Rows}
\usage{
coalesce_dupes(data, ..., pre_sort = TRUE, post_sort = FALSE)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from dbplyr or dtplyr)}

\item{...}{Variables to use for sorting and determining uniqueness. If there
are multiple rows for a given combination of inputs, only the first row
will be preserved. If omitted, simply calls \code{distinct()} with
\code{.keep_all = TRUE}.}

\item{sort}{A logical indicating whether to sort using the input variables.
If no input variables are given, no sorting is performed, and this
parameter is ignored.}
}
\description{
\code{coalesce_dupes()} sorts data, removes duplicates, and combines
information in duplicate rows.
}
\details{
\code{coalesce_dupes()} can be thought of as an enhanced version of
\code{\link[dplyr]{distinct}}. Like \code{distinct()}, \code{coalesce_dupes()}
removes duplicates from a dataset based on a provided set of variables.
Unlike \code{distinct()}, it sorts the data on those variables by default using
\code{\link[dplyr]{arrange}}. It also tries to replace missing values with
the first non-missing values in any duplicate rows using a modification of
\code{\link[dplyr]{coalesce}}.
}
