% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_date.R
\name{coalesce_report_date}
\alias{coalesce_report_date}
\title{Coalesce Report Dates from Archived \code{fst} Files}
\usage{
coalesce_report_date(
  archive_dir = path_create("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/",
    "Report Date Linelist/archive/"),
  id = "inv_local_id"
)
}
\arguments{
\item{archive_dir}{The location of the archive directory}

\item{id}{The name of the column to use when joining data; defaults to
"inv_local_id"}
}
\value{
A \code{tibble} containing the \code{id} column and a \code{report_date} column
}
\description{
\code{coalesce_report_date()} coalesces the output of \code{archive_report_date()} into
a \code{tibble}. Internally, this is a series of join-then-coalesce operations.
}
