% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-covidata.R
\name{coerce_covidata}
\alias{coerce_covidata}
\title{Create \code{as_covidata} Functions}
\usage{
coerce_covidata(class)
}
\arguments{
\item{class}{A new \code{covidata} subclass}
}
\description{
Create a coercion function for a \code{covidata} subclass.
}
\keyword{internal}
