% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cols_exist}
\alias{cols_exist}
\title{Do One or More Columns Actually Exist?}
\usage{
cols_exist(.data, ..., action = rlang::abort)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g. a
\code{\link[tibble:tbl_df-class]{tibble}})}

\item{...}{\code{\link[dplyr:dplyr_tidy_select]{<tidy-select>}} One or more
columns from the data, separated by commas. These can be specified as
unquoted names or using tidy selection helpers; if the latter,
\code{cols_exist()} will check that \emph{at least one} column matches each
provided \code{<tidy-select>} expression.}

\item{action}{The action to perform if one or more columns do not exist; one
of \code{\link[rlang:abort]{abort()}} (the default),
\code{\link[rlang:abort]{warn()}}, or \code{\link[rlang:abort]{inform()}}.}
}
\value{
The input data frame
}
\description{
\code{cols_exist()} is a pipeline validation function that checks whether
specified columns exist in the input data. It is inspired by
\code{pointblank::col_exists()} but allows tidy selection of columns.
}
