% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_xlsx.R
\name{convert_nbs_snapshot}
\alias{convert_nbs_snapshot}
\alias{convert_pcr_snapshot}
\alias{convert_vaccine_snapshot}
\alias{convert_snapshot_to_xlsx}
\title{Find a Snapshot File and Convert to XLSX Format}
\usage{
convert_nbs_snapshot(
  date = Sys.Date(),
  directory = paste0("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/",
    "Sandbox data pull Final/"),
  force = FALSE,
  open = FALSE
)

convert_pcr_snapshot(
  date = Sys.Date(),
  directory = "V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/MSR PCR/",
  force = FALSE,
  open = FALSE
)

convert_vaccine_snapshot(
  date = Sys.Date(),
  directory = path_create("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/",
    "COVID-19 Vaccine Reporting/COVID-19 Vaccine data pull/"),
  force = FALSE,
  open = FALSE
)

convert_snapshot_to_xlsx(
  date = Sys.Date(),
  date_format = "\%Y-\%m-\%d",
  directory,
  force = FALSE,
  open = FALSE
)
}
\arguments{
\item{date}{The date included in the name of the file to convert}

\item{directory}{The directory in which to look for the file to convert}

\item{force}{Should the file be converted if it will overwrite an existing
XLSX file?}

\item{open}{Should the XLSX file be opened after conversion?}

\item{date_format}{The format of the date in the file name. The default for
the generic function is 'YYYY-MM-DD'; see \code{\link[base]{strptime}} for
details on format specification and the body of \code{convert_pcr_snapshot()}
for an implementation example (in 'MMDDYYYY' format).}
}
\value{
A tidy path to the new XLSX file (invisibly)
}
\description{
\code{convert_snapshot_to_xlsx()} finds a snapshot file in \code{directory} with \code{date}
in its name and converts it to an XLSX file of the same name. See
\code{\link{convert_to_xlsx}} for details.
}
\details{
The \verb{convert_*_snapshot()} functions make it easier to convert specific data
files.
}
