% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-covidata.R
\name{covidata-class}
\alias{covidata-class}
\alias{new_covidata}
\alias{validate_covidata}
\title{Construct or Validate \code{covidata} Subclasses}
\usage{
new_covidata(x, ..., date_stamp, nrow, class = NULL)

validate_covidata(x)
}
\arguments{
\item{x}{A tibble-like object.}

\item{...}{Name-value pairs of additional attributes.}

\item{date_stamp}{A \code{\link[base:Dates]{Date}} object indicating the date
that the data was made available}

\item{nrow}{The number of rows, required.}

\item{class}{Subclasses to assign to the new object, default: none.}
}
\description{
Creates or validates a subclass of \code{covidata}, which is just a
\code{\link[tibble:tbl_df-class]{tibble}} with a \code{date_stamp} attribute.

\code{new_covidata()} creates a new object as a subclass of \code{covidata},
\code{tbl_df}, \code{tbl}, and \code{data.frame}. This function is optimized for
performance; checks are reduced to a minimum.

\code{validate_covidata()} checks a \code{covidata} object for internal consistency.
Correct behavior can be guaranteed only if this function runs without raising
an error.
}
\seealso{
\code{\link[tibble]{new_tibble}}
}
