% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-covid_tbl.R
\name{covidata-helpers}
\alias{covidata-helpers}
\alias{ael}
\alias{nbs}
\alias{pcr}
\title{Build a Source-Specific Data Frame}
\usage{
ael(
  ...,
  .date = NULL,
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)

nbs(
  ...,
  .date = NULL,
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)

pcr(
  ...,
  .date = NULL,
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}>
A set of name-value pairs. These arguments are
processed with \code{\link[rlang:nse-defuse]{rlang::quos()}} and support unquote via \code{\link{!!}} and
unquote-splice via \code{\link{!!!}}. Use \verb{:=} to create columns that start with a dot.

Arguments are evaluated sequentially.
You can refer to previously created elements directly or using the \link{.data}
pronoun.
An existing \code{.data} pronoun, provided e.g. inside \code{\link[dplyr:mutate]{dplyr::mutate()}},
is not available.}

\item{.date}{A \code{Date} indicating when the data was made available. This
will usually come from name of the data file; the \code{coviData::load_*}
family detects this automatically when reading the data. This argument is
optional, but omitting it limits the ability of coviData functions to
handle data from several reporting dates.}

\item{.rows}{The number of rows, useful to create a 0-column tibble or
just as an additional check.}

\item{.name_repair}{Treatment of problematic column names:
\itemize{
\item \code{"minimal"}: No name repair or checks, beyond basic existence,
\item \code{"unique"}: Make sure names are unique and not empty,
\item \code{"check_unique"}: (default value), no name repair, but check they are
\code{unique},
\item \code{"universal"}: Make the names \code{unique} and syntactic
\item a function: apply custom name repair (e.g., \code{.name_repair = make.names}
for names in the style of base R).
\item A purrr-style anonymous function, see \code{\link[rlang:as_function]{rlang::as_function()}}
}

This argument is passed on as \code{repair} to \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}.
See there for more details on these terms and the strategies used
to enforce them.}
}
\value{
An object of class \code{ael}, \code{nbs}, or \code{pcr}
}
\description{
\code{ael()}, \code{nbs()}, and \code{pcr()} construct data frames. They are used like
\code{\link[tibble:tibble]{tibble()}}, except an additional \code{.date} argument
is included.
}
\details{
These function create data frames of class \code{ael}, \code{nbs}, or \code{pcr}. The data
behave identically to a \code{tibble} in most functions, but coviData functions
uses these classes to treat each data source appropriately. Additionally, the
\code{.date} argument tells coviData functions when the data was made
available, which facilitates automatic handling of data reported at
different times.
}
\seealso{
\code{\link[tibble:tibble]{tibble()}}
}
