% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coviData-package.R
\docType{package}
\name{coviData-package}
\alias{coviData}
\alias{coviData-package}
\title{coviData: COVID-19 Data Munging Tools for the Shelby County Health Department}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

coviData is designed to help with ETL,
    munging, and basic analysis of COVID-19 data within the Shelby County
    Health Department. It aims to provide a complete framework for
    down/loading, cleaning, and summarizing this data. The goal is to make
    human input easy when creativity is needed and difficult when
    consistency is the priority.
}
\author{
\strong{Maintainer}: Jesse Smith \email{jesse.smith@shelbycountytn.com} (\href{https://orcid.org/0000-0001-7375-2463}{ORCID})

}
\keyword{internal}
