% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-covidata.R
\name{covidata-subclass-creators}
\alias{covidata-subclass-creators}
\alias{new_subclass_fn}
\alias{validate_subclass_fn}
\alias{new_covidata_subclass}
\alias{validate_covidata_subclass}
\title{Create \code{new_subclass()} or \code{validate_subclass()} Functions}
\usage{
new_covidata_subclass(class)

validate_covidata_subclass(class)
}
\arguments{
\item{class}{A new \code{covidata} subclass}
}
\value{
A function for creating or validating a \code{covidata} subclass
}
\description{
\code{new_covidata_subclass()} creates a function for constructing a \code{class}
object.

\code{validate_covidata_subclass()} creates a function for validating a \code{class}
object.

See \code{covidata-subclasses} for more.
}
\keyword{internal}
