% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-covidata.R
\name{covidata-subclasses}
\alias{covidata-subclasses}
\alias{new_ael}
\alias{validate_ael}
\alias{new_nbs}
\alias{validate_nbs}
\alias{new_pcr}
\alias{validate_pcr}
\title{Construct or Validate New Objects of Class \code{ael}, \code{nbs}, or \code{pcr}}
\usage{
new_ael(x, date_stamp, nrow)

new_nbs(x, date_stamp, nrow)

new_pcr(x, date_stamp, nrow)

validate_ael(x)

validate_nbs(x)

validate_pcr(x)
}
\arguments{
\item{x}{A tibble-like object.}

\item{nrow}{The number of rows, required.}
}
\description{
The \code{new_*()} functions create a new object of class \code{ael}, \code{nbs}, or
\code{pcr} as a subclass of \code{covidata}, \code{tbl_df}, \code{tbl}, and \code{data.frame}. They
are optimized for performance; checks are reduced to a minimum.

The \code{validate_*()} functions check an \code{ael}, \code{nbs}, or \code{pcr} object for
internal consistency. Correct behavior can be guaranteed if and only if this
function runs without raising an error.
}
\seealso{
new_covid_tbl validate_covid_tbl
}
\keyword{internal}
