% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-covid_tbl.R
\name{covidata_helper}
\alias{covidata_helper}
\title{Create a Helper Function for a \code{covidata} Subclass}
\usage{
covidata_helper(class)
}
\arguments{
\item{class}{A \code{covidata} subclass}
}
\value{
A user-friendly helper function for constructing objects of \code{class}
}
\description{
\code{covidata_helper} creates helper functions for a \code{covidata} subclass. It
requires \code{new_mysubclass()} and \code{validate_mysubclass()} functions
to be defined and available for the class \code{mysubclass}.
}
\keyword{internal}
