% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_path}
\alias{create_path}
\title{Create a Tidy Path from Directory + File + Extension}
\usage{
create_path(directory, file_name = NULL, ext = NULL)
}
\arguments{
\item{directory}{The path to a directory; optionally, the full path can be
supplied here, in which case \code{create_path()} will simply expand and
tidy the given file path.}

\item{file_name}{The name of the file of interest; may be left \code{NULL} if
this is included in \code{directory}}

\item{ext}{The file extension to add to the path; may be left \code{NULL} if
included in \code{file_name} or \code{directory}}
}
\value{
An \code{fs_path}, which is a character vector with additional attributes
}
\description{
\lifecycle{soft-deprecated}

\code{create_path()} combines a directory, filename, and file extension into
an expanded and tidied path.
}
\details{
\code{create_path()} is now soft-deprecated in favor of
\code{\link[coviData:path_create]{path_create()}}, which is a re-implementation that
follows fs naming conventions, allows both \strong{User}- and \strong{R}-
based home directories, and relies on \code{\link[fs:path]{path()}} to
combine arbitrary path components given in \code{...}.
}
