% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{data-for-regions-snapshots}
\alias{data-for-regions-snapshots}
\alias{download_nbs_snapshot}
\alias{download_pcr_snapshot}
\alias{download_antigen_snapshot}
\alias{download_serology_snapshot}
\title{Download a File from the Data for Regions REDcap Project}
\usage{
download_nbs_snapshot(
  date = Sys.Date(),
  api_token = Sys.getenv("redcap_DFR_token"),
  redcap_file = "nbs_daily_upload",
  directory = paste0("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/",
    "Sandbox data pull Final/"),
  new_file = paste0(date, " Final Data Pull.csv"),
  force = FALSE
)

download_pcr_snapshot(
  date = Sys.Date(),
  api_token = Sys.getenv("redcap_DFR_token"),
  redcap_file = "lab_pcr",
  directory = "V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/MSR PCR/",
  new_file = paste0("MSR - All PCRs_", format(Sys.Date(), "\%m\%d\%Y"), ".csv"),
  force = FALSE
)

download_antigen_snapshot(
  date = Sys.Date(),
  api_token = Sys.getenv("redcap_DFR_token"),
  redcap_file = "lab_antigen",
  directory = "V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/MSR ANTIGEN/",
  new_file = paste0("MSR - All Antigens_", format(Sys.Date(), "\%m\%d\%Y"), ".csv"),
  force = FALSE
)

download_serology_snapshot(
  date = Sys.Date(),
  api_token = Sys.getenv("redcap_DFR_token"),
  redcap_file = "lab_serology",
  directory = "V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/MSR SEROLOGY/",
  new_file = paste0("MSR - All Serologies_", format(Sys.Date(), "\%m\%d\%Y"), ".csv"),
  force = FALSE
)
}
\arguments{
\item{date}{A \code{Date} indicating the date of the file to download}

\item{api_token}{The API token for accessing the Data for Regions REDcap
project. This should be stored in an \emph{.Renviron} file; see
\link{using-renviron} for details.}

\item{redcap_file}{A string indicating the name of the file to download from
REDcap}

\item{directory}{A string specifying the save directory; this should usually
be left alone}

\item{new_file}{A string specifying the save file name; this should always
end in \emph{.csv} and should usually be left alone}

\item{force}{A logical indicating whether to ignore any existing files
matching \code{date} and \code{directory}}
}
\value{
Invisibly returns the path to the new data file
}
\description{
These functions downloads a file posted on the Data for Regions
project. You'll need API access to the project (and an API token for it) to
use this function.

\code{download_nbs_snapshot()} downloads the current snapshot of NBS
investigation data (if posted).

\code{download_pcr_snapshot()} downloads the current snapshot of PCR testing data
(if posted).

\code{download_antigen_snapshot()} downloads the current snapshot of antigen
testing data (if posted).

\code{download_serology_snapshot()} downloads the current snapshot of serology
testing data (if posted).
}
