% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{date-data}
\alias{date-data}
\alias{date_inv}
\alias{date_pcr}
\alias{date_vac}
\title{Get Dates for Data files}
\usage{
date_inv(date = NULL)

date_pcr(date = NULL)

date_vac(date = NULL)
}
\arguments{
\item{date}{A date (as a \code{Date} or \code{"\%Y-\%m-\%d"} string)}
}
\value{
A \code{Date} (or errors if data does not exist)
}
\description{
\code{date_inv()} returns the date of an NBS investigation file (if it exists)
}
\details{
\code{date_pcr()} returns the date of an NBS PCR file (if it exists)

\code{date_vac()} returns the date of a TennIIS vaccination file (if it exists)
}
\keyword{internal}
