% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_date.R
\name{decimal_time}
\alias{decimal_time}
\title{Extract Time as a Decimal of Hours}
\usage{
decimal_time(x)
}
\arguments{
\item{x}{A vector with hour, minute, and/or second information that can be
extracted by lubridate's \code{hour()}, \code{minute()}, and \code{second()} functions
(respectively).}
}
\value{
A \code{double} vector of decimal hours
}
\description{
Extract Time as a Decimal of Hours
}
