% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-helpers.R
\name{distinct_inv}
\alias{distinct_inv}
\title{Get Distinct Investigations}
\usage{
distinct_inv(data = nest_inv(), date = attr(data, "date"), quiet = FALSE)
}
\description{
Get Distinct Investigations
}
