% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_ael}
\alias{download_ael}
\title{Download AEL File from Serv-U SFTP Site}
\usage{
download_ael(
  date = Sys.Date(),
  usr = Sys.getenv("sftp_usr"),
  pwd = Sys.getenv("sftp_pwd"),
  directory = "V:/EPI DATA ANALYTICS TEAM/AEL Data/"
)
}
\description{
\code{download_ael} finds and download the AEL data for the input date and
saves it to a folder of your choice. To make the best use of this function,
you should save your username and password for Serv-U in a ".Renviron" file;
see
}
