% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_and_replace_ael}
\alias{download_and_replace_ael}
\title{Download AEL File from Serv-U SFTP Site}
\usage{
download_and_replace_ael(
  pattern = Sys.Date(),
  usr = Sys.getenv("sftp_usr"),
  pwd = Sys.getenv("sftp_pwd"),
  tofolder = "V:/EPI DATA ANALYTICS TEAM/AEL Data/",
  fix_names = TRUE
)
}
\arguments{
\item{pattern}{A regular expression matching the desired filename in the AEL
folder. The default is todate's date, but a more specific pattern may be
necessary if multiple files are found.}

\item{usr}{Username for the Serv-U site. The default is to look for the
system environment variable "sftp_usr", as storing credentials in a script
is not recommended for security purposes.}

\item{pwd}{Password for the Serv-U site. The default is to look for the
system environment variable "sftp_pwd", as storing credentials in a script
is not recommended for security purposes.}

\item{tofolder}{Folder/directory to save the file to}
}
\description{
Download AEL File from Serv-U SFTP Site
}
