% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_generic.R
\name{download_data_for_regions}
\alias{download_data_for_regions}
\title{Download a File from the Data for Regions REDcap Project}
\usage{
download_data_for_regions(
  date = Sys.Date(),
  api_token,
  redcap_file,
  directory,
  new_file,
  force = FALSE
)
}
\arguments{
\item{date}{A \code{Date} indicating the date of the file to download}

\item{api_token}{The API token for accessing the Data for Regions REDcap
project. This should be stored in an \emph{.Renviron} file; see
\link{using-renviron} for details.}

\item{redcap_file}{A string indicating the name of the file to download from
REDcap}

\item{directory}{A string specifying the save directory; this should usually
be left alone}

\item{new_file}{A string specifying the save file name; this should always
end in \emph{.csv} and should usually be left alone}

\item{force}{A logical indicating whether to ignore any existing files
matching \code{date} and \code{directory}}
}
\value{
Invisibly returns the path to the new data file
}
\description{
\code{download_data_for_regions()} is the workhorse behind the
\verb{download_*_snapshot()} functions.
}
\keyword{internal}
