% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_integrated_data}
\alias{download_integrated_data}
\title{Download Data from the Integrated Data Tool REDcap Project}
\usage{
download_integrated_data(
  date = Sys.Date(),
  api_token = Sys.getenv("redcap_NIT_token"),
  directory = paste0("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/",
    "Integrated data tool Case Interviews/data/"),
  new_file = paste0("integrated_data_", date, ".csv"),
  force = FALSE
)
}
\arguments{
\item{date}{A \code{Date} indicating the date of the file to download}

\item{api_token}{The API token for accessing the Data for Regions REDcap
project. This should be stored in an \emph{.Renviron} file; see
\link{env_vars} for details.}

\item{directory}{A string specifying the save directory; this should usually
be left alone}

\item{new_file}{A string specifying the save file name; this should always
end in \emph{.csv} and should usually be left alone}

\item{force}{A logical indicating whether to ignore any existing files
matching \code{date} and \code{directory}}
}
\value{
Invisibly returns the path to the new data file
}
\description{
\code{download_integrated_data} connects to the Integrated Data Tool REDcap
project via an API token and downloads the data currently available in the
\strong{case interviews} report.
}
\details{
\code{download_integrated_data} downloads data as above and saves to the file
directory and file specified in \code{directory} and \code{new_file} (by
default, it saves to directory
\code{V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/Integrated data
tool Case Interviews/} and file \code{integrated_data_YYYY-MM-DD.csv}.) Only
change this if you're doing some custom analysis; by default,
\code{\link{load_integrated_data}} will look here for data to import.

To use \code{download_integrated_data} you'll need an API access key for the
associated REDcap project. Please contact
\email{jesse.smith@shelbycountytn.gov} or
\email{karim.gilani@shelbycountytn.gov} for more information.

Best practice is to save this API access token in a separate file; see
\link{env_vars} for details.
}
