% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_generic.R
\name{download_interview_report}
\alias{download_interview_report}
\title{Download a Report from the Integrated Data Tool REDcap Project}
\usage{
download_interview_report(
  date = Sys.Date(),
  api_token,
  report_id,
  directory,
  new_file,
  convert = FALSE,
  force = FALSE
)
}
\arguments{
\item{date}{A \code{Date} indicating the date of the file to download}

\item{api_token}{The API token for the REDcap project to access. Ideally,
this should be stored in Renviron and not in a script.}

\item{report_id}{The ID of the report to download}

\item{directory}{A string indicating the directory to save the file in}

\item{new_file}{A string indicating the name to save the file under}

\item{convert}{A logical indicating whether the user would like to convert
the file to another format. If so, the file will open in Excel after
downloading.}

\item{force}{A logical indicating whether to ignore existing files with
the given date already in the directory}
}
\description{
\code{download_interview_report()} downloads a report from the Integrated Data
Tool project. You'll need API access to the project (and an API token for it)
to use this function.
}
\keyword{internal}
