% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_nbs_snapshot}
\alias{download_nbs_snapshot}
\title{Download Snapshot of NBS Data Posted on REDcap}
\usage{
download_nbs_snapshot(
  date = Sys.Date(),
  api_token = Sys.getenv("redcap_DFR_token"),
  redcap_file = "nbs_daily_upload",
  directory = paste0("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/",
    "Sandbox data pull Final/"),
  new_file = paste0(date, " Final Data Pull.csv"),
  convert = FALSE,
  force = FALSE
)
}
\description{
Download Snapshot of NBS Data Posted on REDcap
}
