% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_pcr_snapshot}
\alias{download_pcr_snapshot}
\title{Download Snapshot of NBS PCR Test Data Posted on REDcap}
\usage{
download_pcr_snapshot(
  date = Sys.Date(),
  api_token = Sys.getenv("redcap_DFR_token"),
  redcap_file = "lab_pcr",
  directory = "V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/MSR PCR/",
  new_file = paste0("MSR - All PCRs_", format(Sys.Date(), "\%m\%d\%Y"), ".csv"),
  convert = FALSE,
  force = FALSE
)
}
\description{
Download Snapshot of NBS PCR Test Data Posted on REDcap
}
