% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_redcap_file}
\alias{download_redcap_file}
\title{Download REDcap File}
\usage{
download_redcap_file(
  date = Sys.Date(),
  api_token,
  redcap_file,
  directory,
  new_file,
  convert = FALSE,
  force = FALSE
)
}
\description{
Download REDcap File
}
