% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_state_contacts.R
\name{download_state_contacts}
\alias{download_state_contacts}
\title{Download "Automated Contacts for State Monitoring" Report}
\usage{
download_state_contacts(
  date = Sys.Date(),
  api_token = Sys.getenv("redcap_NIT_token"),
  directory = "V:/State Contact Monitoring Spreadsheets/REDcap Exports",
  report_id = "176",
  force = FALSE
)
}
\arguments{
\item{date}{The date to use in the file name}

\item{api_token}{API token for integrated data tool}

\item{directory}{Directory to save under; should usually be left alone.}

\item{force}{Should the new file overwrite any existing files?}
}
\description{
\code{download_state_contacts()} saves the
"Automated Contacts for State Monitoring" report for transformation and
upload to the Tennessee DoH REDcap.
}
