% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_date.R
\name{dttm_to_dt}
\alias{dttm_to_dt}
\title{Coerce Datetimes to Dates if No Information is Lost}
\usage{
dttm_to_dt(.x, force = c("none", "dt", "dttm"))
}
\arguments{
\item{.x}{A \code{Date}, \code{POSIXct} or \code{POSIXlt} (i.e. a datetime) vector}

\item{force}{Whether to force conversion to \code{Date} (\code{force = "dt"}) or
\code{POSIXct} (\code{force = "dttm"}). The default is no forcing (\code{force = "none"}).}
}
\value{
Either a \code{POSIXct} vector or a \code{Date} vector
}
\description{
\code{dttm_to_dt()} converts \code{POSIXt} objects to \code{Date} objects when there is no
additional information contained in the \code{POSIXt} format (i.e. there is no
sub-daily information).
}
\details{
Specifically, \code{dttm_to_dt} checks whether all sub-daily information is the
same for each value in the datetime vector. If so, no additional information
is gained by using a \code{POSIXt} format over the simpler \code{Date} format, and
the data is coerced.

If the input is scalar (i.e. has length \code{1L}), then no conversion is
attempted
}
