% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{error_notify}
\alias{error_notify}
\title{Send an Email Notification on Error}
\usage{
error_notify(
  abort = TRUE,
  to = "jesse.smith@shelbycountytn.gov",
  operation = "An Operation",
  subject = paste(operation, "failed at", format(Sys.time(), "\%H:\%M on \%Y-\%m-\%d")),
  body = NULL,
  cc = NULL,
  html = FALSE
)
}
\arguments{
\item{abort}{Should execution terminate after reporting the error?}

\item{to}{A character vector containing recipient email addresses. Required.}

\item{subject}{A character string containing the email subject. Required.}

\item{body}{A character string containing the email body; if \code{html = TRUE},
this will be parsed as HTML. Optional.}

\item{cc}{A character vector containing email address to "cc". Optional.}

\item{html}{Should the \code{body} be treated as plain text or HTML? The default
is plain text (\code{FALSE}).}
}
\value{
The input error condition, with a backtrace
}
\description{
\code{error_notify()} generates an error handler that sends an email notification
when an expression generates an error. It relies on
\code{\link[coviData:notify]{notify()}} and can optionally avoid terminating
the larger expression it was called from. This is useful for running scripts
with independent components.
}
\keyword{internal}
