% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_dates}
\alias{fill_dates}
\title{Fill Implicitly Missing Dates in a Time-Indexed Data Frame}
\usage{
fill_dates(data, date_col, start = NULL, end = NULL, fill_na = NA)
}
\arguments{
\item{data}{A data frame or data frame extension (e.g. a tibble)}

\item{date_col}{The column to use as the time index}

\item{start}{An optional start date for the period to fill}

\item{end}{An optional end date for the period to fill}

\item{fill_na}{The value to use when creating observations for implicitly
missing data}
}
\value{
A tibble
}
\description{
\code{fill_date()} finds implicitly missing observations in a time-indexed data
frame and fills them with \code{fill_na}. \code{start} and \code{end} values can be supplied
to fill dates outside the range of the data.
}
\keyword{internal}
