% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{filter_region}
\alias{filter_region}
\title{Filter Dataframe from an AEL File to Memphis Area}
\usage{
filter_region(
  .data,
  states = c("TN"),
  zips = c("380", "381"),
  incl_msr = FALSE,
  incl_na = TRUE
)
}
\arguments{
\item{.data}{A data frame or data frame extention (e.g. a tibble)}

\item{states}{Two-letter codes for the states to include in the output}

\item{zips}{1-to-5 digit codes matching the ZIP codes to include in the
output}

\item{incl_msr}{A logical indicating whether to ensure all states/ZIPs in
the Memphis Statistical Region are included}

\item{incl_na}{A logical indicating whether missing values are included or
excluded}
}
\value{
A data frame, tibble, etc. containing observations from the specified
region
}
\description{
\code{filter_region} takes a data frame from \code{link{read_ael}} & removes
observations not in the region specified by \code{states} and \code{zips}.
The default is to keep Shelby County residents and missing values;
\code{incl_msr} ensures that the entire Memphis Statistical Region is
included (as well as any other regions specified). Note that \code{zips}
match to starting values of \code{PtZipcode}; if a 5-digit code is specified,
this is equivalent to matching an entire ZIP code.
}
