% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{fst_convert}
\alias{fst_convert}
\title{Convert Tabular Data Files to \code{fst} Format}
\usage{
fst_convert(path, save_as = NULL, ...)
}
\arguments{
\item{path}{Character. File path to read from. it will be passed to
\code{\link[coviData:read_file_excel]{read_file_excel()}} if the file
extension is "xlsx" or "xls", otherwise it will be passed to
\code{\link[coviData:read_file_delim]{read_file_delim()}}}

\item{save_as}{Character. The path to save the result \code{fst} file. This must
have an \code{fst} extension; if it does not, the current extension (if any)
will be removed and replaced with \code{fst}. If left \code{NULL}, the \code{path}
argument is used (with extension replaced by \code{fst}).}

\item{...}{Additional arguments to pass to \code{read_file_delim()} or
\code{read_file_excel()}. The \code{col_types} argument is particularly useful to
control the type in which the data is saved (default is all \code{character}).}
}
\value{
The read \code{data}, invisibly
}
\description{
\code{fst_convert()} reads Excel, delimited, and fixed-width data file and
converts them to \code{fst} format.
}
