% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gg_var}
\alias{gg_var}
\title{Extract an Axis Variable Name}
\usage{
gg_var(gg_obj, axis = c("x", "y"))
}
\arguments{
\item{gg_obj}{A \code{ggplot} object}

\item{axis}{The axis variable to extract}
}
\value{
A character vector of length 1 containing the variable name
}
\description{
\code{gg_var()} extracts the name of the variable used as the \code{"x"} or \code{"y"}
\code{axis} in a \code{ggplot} object.
}
\keyword{internal}
