% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_open}
\alias{is_open}
\title{Test Whether a File is Open}
\usage{
is_open(path)
}
\arguments{
\item{path}{The path to the file in question}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
\code{is_open()} tests whether access to a file is currently being blocked by
another program. TODO: This should probably be named \code{is_blocked()}.
}
\keyword{internal}
