% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_windows}
\alias{is_windows}
\title{Check Whether \strong{R} is Running in Windows}
\usage{
is_windows()
}
\description{
\code{is_windows()} checks whether \strong{R} is running on Windows. There are
several functions in coviData that will only work on Windows.
}
\keyword{internal}
