% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_and_process_ael}
\alias{load_and_process_ael}
\title{Load, Clean, & Merge AEL Data Files for 2 Consecutive dates}
\usage{
load_and_process_ael(
  date1 = Sys.Date(),
  directory = "V:/EPI DATA ANALYTICS TEAM/AEL Data/",
  date2 = date1 - 1,
  date1_file = NULL,
  date2_file = NULL,
  overwrite_names_1 = FALSE,
  overwrite_names_2 = FALSE,
  encoding = TRUE,
  check = TRUE
)
}
\description{
DEPRECATED - please use \code{load_ael() \%>\% clean_generic()}.
}
