% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_sas}
\alias{load_sas}
\title{Load Cleaned Data from SAS Program}
\usage{
load_sas(
  directory = "V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/Data for R/",
  dataset = c("positive_ppl", "negative_ppl", "positive_pcr", "negative_pcr"),
  ext = c("csv", "xlsx")
)
}
\arguments{
\item{directory}{A string indicating the directory of the file to read}

\item{dataset}{A string indicating which dataset to read}

\item{ext}{The file extension to search for - "csv" by default}
}
\description{
Load Cleaned Data from SAS Program
}
