% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_end}
\alias{log_end}
\title{End a Logging Session}
\usage{
log_end(sink = getOption(".log_sink_"), con = getOption(".log_connection_"))
}
\arguments{
\item{sink}{The message sink used to divert stderr output to the log file}

\item{con}{The open file connection to the log file}
}
\value{
The path to the log file (invisibly)
}
\description{
\code{log_end()} ends a logging session by closing both the message sink and the
file connection to the log file.
}
