% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_start}
\alias{log_start}
\title{Start a Logging Session}
\usage{
log_start(file, dir = log_dir(), append = TRUE, force = FALSE)
}
\arguments{
\item{file}{The log file name. If path components are included they are
assumed to be relative to \code{dir}.}

\item{dir}{The log directory. \code{log_dir()}, the default, returns the current
log directory, or \code{NULL} if none is set. The latter case will cause \code{file}
to be created in or relative to your current working directory.}

\item{append}{Should an existing file be appended to?}

\item{force}{If \code{append = FALSE}, should an existing file be overwritten?}
}
\value{
The path to the log file (invisibly)
}
\description{
\code{log_start()} begins diverting message stream (stderr) output to the
specified log file. Only one logging session can be active at a time.
}
