% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{notify}
\alias{notify}
\title{Send an Email Notification with Outlook}
\usage{
notify(
  to,
  subject,
  body = NULL,
  cc = NULL,
  html = FALSE,
  importance = c("Normal", "Low", "High")
)
}
\arguments{
\item{to}{A character vector containing recipient email addresses. Required.}

\item{subject}{A character string containing the email subject. Required.}

\item{body}{A character string containing the email body; if \code{html = TRUE},
this will be parsed as HTML. Optional.}

\item{cc}{A character vector containing email address to "cc". Optional.}

\item{html}{Should the \code{body} be treated as plain text or HTML? The default
is plain text (\code{FALSE}).}

\item{importance}{The importance (priority, etc) to attach to the email. The
default is \code{"Normal"}; alternatives are \code{"Low"} or \code{"High"}.}
}
\description{
\code{notify} sends a notification email via Outlook. This assumes that Outlook
is installed on your machine and that you've already set up an account.
}
\keyword{internal}
