% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_by_date.R
\name{path_by_date}
\alias{path_by_date}
\alias{path_inv}
\alias{path_pcr}
\alias{path_vac}
\title{Return the Path to a NBS Investigations File}
\usage{
path_by_date(
  dir,
  date_format,
  date_regex,
  date = NULL,
  file_regex = ".*{date}.*",
  type = "any",
  force_latest = TRUE
)

path_inv(
  date = NULL,
  dir = path_create("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/",
    "Sandbox data pull Final/"),
  ext = c("csv", "xlsx")
)

path_pcr(
  date = NULL,
  dir = "V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/MSR PCR/",
  ext = c("csv", "xlsx")
)

path_vac(
  date = NULL,
  dir = path_create("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/",
    "COVID-19 Vaccine Reporting/data/COVID-19 vaccine data pull/"),
  ext = c("csv", "xlsx")
)
}
\arguments{
\item{dir}{The directory holding the investigations files}

\item{date_format}{The format of the date in the file name; see
\code{\link[base:format]{format()}} for more information}

\item{date_regex}{A regular expression matching the date in the file name.
This will hopefully be deprecated in the future, but is currently needed
for extracting dates from the file path strings.}

\item{date}{A \code{Date} or string in the format "YYYY-MM-DD"}

\item{file_regex}{A regular expression matching the file names of interest.
The location of the date in the file name is specified using \code{"{date}"}, as
in the glue package.}

\item{type}{The file type(s) to return, one or more of \code{"any"}, \code{"file"},
\code{"directory"}, \code{"symlink"}, \code{"FIFO"}, \code{"socket"}, \code{"character_device"} or
\code{"block_device"}.}

\item{force_latest}{If multiple files with the given \code{date} are found,
should the function return only the file with the latest creation date
(\code{TRUE}, the default), or should it return all file paths (\code{FALSE})?}

\item{ext}{The file type to return. NBS investigations files are saved as
both "xlsx" and "csv" files.}
}
\value{
An \code{fs_path} character vector
}
\description{
\verb{path_*} functions return paths to one or more files of interest
containing the supplied date in their names and ending with the supplied
file extension. These are convenience wrappers around \code{path_by_date()}, which
returns path that contain a specified date given additional parameters.
}
\details{
\itemize{
\item{\code{path_inv()} returns path(s) to NBS investigations files}
\item{\code{path_pcr()} returns path(s) to NBS PCR files}
\item{\code{path_vac()} returns path(s) to TennIIS vaccination files}
}

This will eventually replace \code{\link[coviData:find_file]{find_file()}}
for finding NBS investigations files.
}
