% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_by_pattern.R
\name{path_deaths}
\alias{path_deaths}
\title{Most Recent Deaths File}
\usage{
path_deaths(
  src = c("nbs", "surv"),
  dir = "V:/EPI DATA ANALYTICS TEAM/MORTALITY DATA/"
)
}
\arguments{
\item{src}{Which deaths file to look for; \code{"nbs"} for the NBS file, or
\code{"surv"} for the surviellance file}

\item{dir}{The directory to search}
}
\value{
An \code{fs_path} character vector
}
\description{
\code{path_deaths()} finds the path to the most recent linelist of deaths from
NBS or the surveillance team.
}
