% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_inv.R
\name{path_inv}
\alias{path_inv}
\title{Return the Path to a NBS Investigations File}
\usage{
path_inv(
  date = NULL,
  dir = path_create("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/",
    "Sandbox data pull Final/"),
  ext = c("csv", "xlsx")
)
}
\arguments{
\item{date}{A \code{Date} or string in the format "YYYY-MM-DD"}

\item{dir}{The directory holding the investigations files}

\item{ext}{The file type to return. NBS investigations files are saved as
both "xlsx" and "csv" files.}
}
\value{
An \code{fs_path} character vector
}
\description{
\code{path_inv()} returns paths to one or more NBS investigations files containing
the supplied date in their names and ending with the supplied file extension.
}
\details{
This will eventually replace \code{\link[coviData:find_file]{find_file()}}
for finding NBS investigations files.
}
