% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_by_date.R
\name{path_inv_id}
\alias{path_inv_id}
\title{Path to Processed Investigation Local IDs}
\usage{
path_inv_id(
  date = NULL,
  status = c("+", "-"),
  dir = path_create("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/Data for R/",
    "inv_local_id")
)
}
\arguments{
\item{date}{A \code{Date} or string in the format "YYYY-MM-DD"}

\item{status}{Should the path be to \code{"positive"} or \code{"negative"} IDs?}

\item{dir}{The directory holding the files of interest}
}
\value{
An \code{fs_path}
}
\description{
Path to Processed Investigation Local IDs
}
\keyword{internal}
