% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_generic.R
\name{preprocess}
\alias{preprocess}
\title{Standardize Variables, Handle Dates, and Remove Bad Columns}
\usage{
preprocess(data)
}
\arguments{
\item{data}{A dataframe(-like) object}
}
\value{
A processed tibble
}
\description{
\code{preprocess} standardizes column names, converts variables to the
appropriate data types (including oddly formatted dates/datetimes), and
removes empty or constant columns.
}
