% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process-nbs}
\alias{process-nbs}
\alias{process_inv}
\alias{process_pcr}
\alias{pos}
\alias{neg}
\alias{process_positive_people}
\alias{process_negative_people}
\alias{process_positive_tests}
\alias{process_negative_tests}
\title{Run the Cases Workflow for Daily Report}
\usage{
process_inv(
  data = read_inv(),
  date = attr(data, "date"),
  distinct = TRUE,
  replace = FALSE,
  quiet = FALSE
)

process_pcr(
  data = read_pcr(),
  inv = process_inv(read_inv(date = date)),
  date = attr(data, "date"),
  quiet = FALSE
)

pos(data)

neg(data)

process_positive_people(data = read_inv(date), date = NULL, replace = FALSE)

process_negative_people(data = read_inv(date), date = NULL, replace = FALSE)

process_positive_tests(data = read_pcr(date), date = NULL)

process_negative_tests(data = read_pcr(date), date = NULL)
}
\arguments{
\item{data}{A \code{tibble}. Input data- either investigations for
\code{process_inv()} or PCR tests for \code{process_pcr()}}

\item{date}{The download date of the data. If the data has a \code{date}
attribute, this is used; otherwise, the date must be supplied. See
\code{Details} above.}

\item{replace}{Should the filtering results be recalculated and replace the
existing results (if any)?}

\item{quiet}{Should intermediate step summaries be output as messages?}

\item{inv}{A \code{tibble}. Input investigation data for \code{process_pcr()}, as
returned by \code{process_inv()}.}
}
\value{
A nested 2x2 \code{tibble} with \code{logical} column \code{positive} and
\verb{list_of<tbl_df>} column \code{data}; the latter contains the transformed
positive and negative data
}
\description{
The \verb{process_*} functions transform NBS data in preparation for analysis.

\code{process_inv()} transforms investigation data.

\code{process_pcr()} transforms PCR data.

\code{pos()} and \code{neg()} extract positive or negative data from the result.
}
\details{
\code{process_inv()} is special in that it saves records identified as positive
and negative to the \verb{V:} drive as an \code{fst} file. This allows significant
speedup on subsequent runs, since the only necessary operations are mutating
of key variables prior to filtering.

\code{process_pcr()} joins PCR data to the results of \code{process_inv()}. These
results can be supplied by the \code{inv} argument; by default, they are
calculated on-the-fly.

Both functions need the download date of their data.
\code{\link[coviData:read_inv]{read_inv()}} and
\code{\link[coviData:read_pcr]{read_pcr()}} supply this as a \code{date}
attribute; if this exists in the input data, it is used
(with a warning if it conflicts with the \code{date} argument). If it does not
exist, the \code{date} argument defaults to the most recent download date;
in general, the \code{date} should either be supplied by an attribute (as above)
or explicitly passed to the \code{date} argument.

\code{process_positive_people()}, \code{process_negative_people()},
\code{process_positive_tests()}, and \code{process_negative_tests()} are deprecated
in favor of \code{pos(process_inv())}, \code{neg(process_inv())}, \code{pos(process_pcr())},
and \code{neg(process_pcr())}
}
