% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_names}
\alias{process_names}
\title{Add Name Columns to AEL Excel File}
\usage{
process_names(data, force = FALSE)
}
\arguments{
\item{data}{A data frame or tibble containing AEL data from one date}

\item{force}{A logical; if \code{force = TRUE}, the data is (re-)processed
regardless of whether it passes the above checks}
}
\value{
A tibble containing the input data plus standardized
\code{PatientLastName}, \code{PatientFirstName}, \code{PatientMiddleName},
and \code{PatientMidOthName} columns
}
\description{
\code{process_names} creates standardized \code{PatientLastName},
\code{PatientFirstName}, \code{PatientMiddleName}, and
\code{PatientMidOthName} columns in AEL data, if they do not already exist.
}
\details{
\code{process_names} first checks the input data for the above columns, as
well as any additional \code{PatientMidOthName} columns. If any of the
standardized columns do not exist, or non-standard \code{Patient...Name}
columns exist, the function deletes any previously created
\code{Patient...Name} columns and adds standardized ones using
\code{\link{split_names}}. If the data contains the standardized columns and
no additional \code{PatientMidOthName} columns, the data is returned as-is.
}
