% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_nbs}
\alias{process_nbs}
\alias{process_positive_people}
\alias{process_negative_people}
\alias{process_positive_tests}
\alias{process_negative_tests}
\title{Run the Cases Workflow for Daily Report}
\usage{
process_positive_people(data = read_inv(date = date), date = NULL)

process_negative_people(data = read_inv(date = date), date = NULL)

process_positive_tests(data = read_pcr(date = date), date = NULL)

process_negative_tests(data = read_pcr(date = date), date = NULL)
}
\arguments{
\item{data}{A \code{tibble}. Input data- either investigations for
\verb{process_*_people()} functions or PCR tests for \verb{process_*_tests()}
functions}

\item{date}{The download date of the data file to use if no data is supplied;
defaults to most recent file}
}
\value{
A \code{tibble} containing the information from the NBS investigations/PCR
files on positive/negative people/tests
}
\description{
The \verb{process_*} functions mimic data transformations in the SAS program for
positive/negative people/tests.
}
