% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_positive_people}
\alias{process_positive_people}
\title{Run the Cases Workflow for Daily Report}
\usage{
process_positive_people(date = NULL)
}
\arguments{
\item{date}{The date of the files to use}
}
\value{
A \code{tibble} containing the information from the NBS investigations
file on all cases
}
\description{
\code{process_positive_people()} runs the positive people (cases) workflow from
the SAS program.
}
