% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{read_ael}
\alias{read_ael}
\title{Read AEL Data from Excel File}
\usage{
read_ael(path)
}
\arguments{
\item{path}{The file path to the Excel file}
}
\value{
A \code{\link[tibble]{tibble}} containing the contents of the Excel
file
}
\description{
\code{read_xl} reads in an xls(x) file, converts date columns to
\code{Date} class, and converts \code{character} columns to \code{factor}
class. \code{read_ael} uses all columns of a file to guess the column type.
It is intended for reading an Excel file containing data sent from AEL to the
Shelby County Health Department.
}
