% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_inv_id}
\alias{read_inv_id}
\title{Read Investigation ID by Status}
\usage{
read_inv_id(date = NULL, status = c("+", "-"))
}
\arguments{
\item{date}{The download date of the data}

\item{status}{Should positive or negative IDs be returned?}
}
\value{
A \code{tibble} with \code{character} column \code{inv_local_id}
}
\description{
Read Investigation ID by Status
}
\keyword{internal}
