% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_vac}
\alias{read_vac}
\title{Read Vaccination Data}
\usage{
read_vac(
  date = NULL,
  col_select = NULL,
  col_types = vroom::cols(.default = vroom::col_character()),
  ...
)
}
\arguments{
\item{date}{The download date of the data}

\item{col_select}{One or more selection expressions, like in
\code{dplyr::select()}. Use \code{c()} or \code{list()} to use more than one expression.
See \code{?dplyr::select} for details on available selection options.}

\item{col_types}{One of \code{NULL}, a \code{\link[readr:cols]{cols()}} specification, or
a string. See \code{vignette("readr")} for more details.

If \code{NULL}, all column types will be imputed from the first 1000 rows
on the input. This is convenient (and fast), but not robust. If the
imputation fails, you'll need to supply the correct types yourself.

If a column specification created by \code{\link[readr:cols]{cols()}}, it must contain
one column specification for each column. If you only want to read a
subset of the columns, use \code{\link[readr:cols]{cols_only()}}.

Alternatively, you can use a compact string representation where each
character represents one column:
c = character, i = integer, n = number, d = double,
l = logical, f = factor, D = date, T = date time, t = time, ? = guess, or
\code{_}/\code{-} to skip the column.}

\item{...}{Additional arguments to pass to
\code{\link[coviData:read_file_delim]{read_file_delim()}}}
}
\value{
A \code{tibble}
}
\description{
Read Vaccination Data
}
