% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_report_date.R
\name{reduce_fst_report_dates}
\alias{reduce_fst_report_dates}
\title{Join Report Dates from \code{fst} Files}
\usage{
reduce_fst_report_dates(.accumulated, .next, .by, .pb = NULL)
}
\arguments{
\item{.accumulated}{The accumulated results from \code{reduce()}}

\item{.next}{The next data set to reduce}

\item{.by}{The name of the column to use when joining data; defaults to
"inv_local_id"}

\item{.pb}{An optional progress bar object created by
\code{\link[progress:progress_bar]{progress_bar()}}; will be incremented
after each reduction step}
}
\value{
The reduced and coalesced data as a \code{tibble}
}
\description{
\code{reduce_fst_report_dates()} is the workhorse of \code{coalesce_report_date()}. It
performs the \code{reduce()} operations that sequentially convert fst objects to
tibbles, left join the result, and coalesce the \code{report_date} columns in each
to a single column. The column from \code{.next} gets preference in the
\code{coalesce()} operation.
}
\keyword{internal}
