% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace_ael}
\alias{replace_ael}
\title{Add Updated Names Columns to AEL File and Replace Existing File}
\usage{
replace_ael(
  date = Sys.Date(),
  directory = "V:/EPI DATA ANALYTICS TEAM/AEL Data/",
  date_file = NULL,
  overwrite_names = FALSE
)
}
\arguments{
\item{date}{The date of the file to replace}

\item{directory}{The directory in which the file is located}

\item{date_file}{IDK}

\item{overwrite_names}{If names are already cleaned, should this be
overwritten?}
}
\description{
Function to replace AEL data with cleaned names data.
}
