% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace_deaths_id}
\alias{replace_deaths_id}
\title{Replace NBS ID in Surveillance Deaths Linelist with Standardized ID}
\usage{
replace_deaths_id(
  file = "Working Copy Death  Epi.xlsx",
  directory = "V:/EPI DATA ANALYTICS TEAM/MORTALITY DATA/",
  id = NULL,
  save_as = paste0("sas_data/cleaned_surveillance_copy.xlsx")
)
}
\arguments{
\item{file}{The name of the surveillance linelist file}

\item{directory}{The directory containing the linelist}

\item{id}{The name of the column specifying the NBS ID in the linelist; if
\code{NULL}, assumes that this is the first column}

\item{save_as}{The file path specifying where to save the results}
}
\value{
The cleaned data as a \code{\link[tibble]{tibble}} (invisibly)
}
\description{
\code{replace_deaths_id()} facilitates comparison of the surveillance deaths
linelist with the one exported from NBS by standardizing patient IDs. It
saves the result in the file specified in \code{save_as}.
}
