% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{sample_unique}
\alias{sample_unique}
\title{Sample from Unique Values of Data}
\usage{
sample_unique(.x, n)
}
\arguments{
\item{.x}{The data to sample}

\item{n}{The number of samples to return}
}
\value{
The shuffled input and additional observations
}
\description{
\code{sample_unique()} samples from the unique values in the input. If
\code{size} is greater than or equal to the number of input observations, it
ensures that each observations is present at least once. All sampling steps
are uniform random, but the output will only be uniform random for \code{n} <=
number of observations, or \code{n} >> number of observations. This is due to the
switch from sampling with replacement to sampling without replacement once
the sample size is larger than the number of observations.
}
