% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{save_plot}
\alias{save_plot}
\title{Save a \code{ggplot} Object}
\usage{
save_plot(
  gg_obj = ggplot2::last_plot(),
  path,
  ratio = c(16, 9),
  size = 0.75,
  ...
)
}
\arguments{
\item{gg_obj}{A \code{ggplot} object}

\item{path}{The location and file for saving the plot. The image type is
inferred from the file extension.}

\item{ratio}{The aspect ratio of the figure; defaults to \code{c(16, 9)}}

\item{size}{The relative size of the figure; defaults to \code{0.75}, meaning that
the figure width will be \code{0.75 * ratio[[1]]}, and the figure height will be
\code{0.75 * ratio[[2]]}. Useful for scaling an aspect ratio without calculating
the precise dimensions.}

\item{...}{Additional arguments to pass to
\code{\link[ggplot2:ggsave]{ggsave()}}}
}
\value{
The \code{ggplot} object (invisibly)
}
\description{
\code{save_plot()} saves a \code{ggplot} object to the specified path.
}
