% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_date.R
\name{save_report_date}
\alias{save_report_date}
\title{Save Report Dates to \code{fst} File}
\usage{
save_report_date(
  .data,
  save_as = path_create("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/",
    "Report Date Linelist/report_date_linelist", ext = "fst")
)
}
\arguments{
\item{.data}{A \code{tibble} returned from \code{coalesce_report_date()} or
\code{add_collection_date()}}

\item{save_as}{The write path to the new \code{fst} file}
}
\value{
The input data (invisibly)
}
\description{
\code{save_report_date()} is a simple wrapper around
\code{\link[fst:write_fst]{write_fst()}} that saves reduced report date data
to an \code{fst} file. It exists purely for convenience and semantic clarity; it
can only be used with the output of \code{coalesce_report_date()} or
\code{add_collection_date()}.
}
