% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_calendar.R
\name{schedule_calendar}
\alias{schedule_calendar}
\title{Create a Scheduling Calendar}
\usage{
schedule_calendar(
  title = "Work Schedule",
  schedule = c("weekdays", "42", "5623", "custom"),
  start = "2021-01-01",
  end = "2021-12-31",
  anchor = start,
  cycle = NULL,
  .pdf = FALSE,
  .dir = "~/"
)
}
\arguments{
\item{title}{The title of the calendar}

\item{schedule}{The schedule to use}

\item{start}{The start date of the returned schedule; either a string in
"YYYY-MM-DD" format or a \code{Date} object}

\item{end}{The end date of the returned schedule; either a string in
"YYYY-MM-DD" format or a \code{Date} object}

\item{anchor}{The date from which to start ("anchor") schedule calculations.
This can be any valid date; no particular relationship to \code{start} or \code{end}
is needed. It must be either a string in "YYYY-MM-DD" format or a
\code{Date} object.}

\item{cycle}{A logical vector defining one scheduling cycle. If named, names
are passed to \code{\link[coviData:parse_weekday]{parse_weekday()}} for
standardization of weekday names.}

\item{.pdf}{Should the schedule be saved as a pdf?}

\item{.dir}{The directory in which to save the schedule if \code{.pdf = TRUE};
defaults to the \code{R_USER} directory}
}
\value{
A list defining the \code{calendR} object
}
\description{
\code{schedule_calendar()} creates a calendar from a given \code{schedule}, \code{start}
date, \code{end} date, and (if needed) \code{anchor} date. Custom schedules can be used
by setting \code{schedule = "custom"} and providing the scheduling cycle via
\code{cycle}. The schedule can be saved to pdf with \code{.pdf = TRUE}; the file name
is a standardized version of the title.
}
