% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_colnames}
\alias{select_colnames}
\title{Select Column Names Using Tidyselect Specifications}
\usage{
select_colnames(.data, ...)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g. a \code{tibble})}

\item{...}{\verb{<tidy-select>} One or more tidyselect specifications for the
desired column names (including simply using those column names)}
}
\value{
A character vector of column names
}
\description{
\code{select_colnames()} selects the names of columns specified in \code{...}. It is
useful for standardizing a function's interface while providing a link to
underlying functions that may take a variety of column specifications.
}
\keyword{internal}
