% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{set_axis_limits}
\alias{set_axis_limits}
\title{Set Viewing Limits on a Plot}
\usage{
set_axis_limits(gg_obj, xlim = NULL, ylim = NULL)
}
\arguments{
\item{gg_obj}{A \code{ggplot} object}

\item{xlim}{The limits for the x axis, in the scale of the x axis}

\item{ylim}{The limits for the y axis, in the scale of the y axis}
}
\value{
The modified \code{ggplot} object
}
\description{
\code{set_limits()} sets limits in the viewing panel of a \code{ggplot} object.
}
\keyword{internal}
