% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{set_covid_theme}
\alias{set_covid_theme}
\title{Set a Standard Theme for SCHD COVID Plots}
\usage{
set_covid_theme(gg_obj)
}
\arguments{
\item{gg_obj}{A \code{ggplot} object}
}
\value{
The modified \code{ggplot} object
}
\description{
\code{set_covid_theme()} is a wrapper around
\code{\link[ggthemes:theme_fivethirtyeight]{theme_fivethirtyeight(base_size = 14)}}.
It standardizes the base theme used by all SCHD COVID plots.
}
\keyword{internal}
