% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp.R
\name{sftp_changedir}
\alias{sftp_changedir}
\title{Move to another directory in an SFTP account}
\usage{
sftp_changedir(
  tofolder,
  current_connection_name = "sftp_con",
  verbose = TRUE,
  curlPerformVerbose = FALSE
)
}
\arguments{
\item{tofolder}{A character vector of length 1 containing the name of a
folder to where you want to go. The folder(s) will be looked for below the
folder where you are currently standing according to the sftp connection
list object you are using (default 'stfp_con'). A subfolder (or a
sub-subfolder, etc.) can be specified in one single operation by supplying
a path, like "folder1/folder2/folder3". To move up one level in the folder
hierarchy, use \code{"../"}. To go to the root folder, use "root".}

\item{current_connection_name}{A string giving the name of the SFTP
connection list object currently used, created  by calling
\code{\link{sftp_connect}}. Default is \code{"sftp_con"}.}

\item{verbose}{Logical. Turn on messages to console. Default is TRUE}

\item{curlPerformVerbose}{Logical. Turn on messages to console form
curlPerform. Default is FALSE.}
}
\value{
The function returns TRUE after successful change of directory.
}
\description{
Change your "current directory" used by the sftp functions.
The function uses and modifies the connection credentials in a list object
created by calling \link{sftp_connect}.
}
\examples{

\dontrun{
# minimal - move up one level, rely on defaults
sftp_changedir("..")

# minimal - to to root, rely on defaults
sftp_changedir("root")

# explicit - change to a directory at the end of a multi-level path
sftp_changedir(tofolder = "level1/folder2/directory3",
               current_connection_name = "sftp_con",
               verbose = TRUE,
               curlPerformVerbose = FALSE)
}

}
\seealso{
\link{sftp_makedir}, \link{sftp_removedir}, \link{sftp_rename}

Other sftp: 
\code{\link{sftp_connect}()},
\code{\link{sftp_delete}()},
\code{\link{sftp_download}()},
\code{\link{sftp_listdirs}()},
\code{\link{sftp_listfiles}()},
\code{\link{sftp_list}()},
\code{\link{sftp_makedir}()},
\code{\link{sftp_removedir}()},
\code{\link{sftp_rename}()},
\code{\link{sftp_upload}()}
}
\author{
\href{https://github.com/stenevang/sftp}{Theodor Stenevang Klemming}
}
\concept{sftp}
\keyword{internal}
