% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp.R
\name{sftp_connect}
\alias{sftp_connect}
\title{Create a List Object with SFTP Connection Details}
\usage{
sftp_connect(
  server = "",
  folder = "",
  username = "",
  password = "",
  protocol = "sftp://",
  port = 22
)
}
\arguments{
\item{server}{The SFTP server URL, hostname, or IP address}

\item{folder}{The path to the folder where you want to operate; can be a
single folder name in the SFTP root or a path with subfolders, e.g.
\code{dir1/dir2/dir3}. Note that directory names are case sensitive.
Default is empty string, which takes you to the root folder.}

\item{username}{The SFTP account username. It is recommended to not store
your username in any script, but instead create an snvironment variable and
call it using \code{\link[base]{Sys.getenv}}.}

\item{password}{The SFTP account password. It is recommended to not store
your password in any script, but instead create an environment variable and
call it using \code{\link[base]{Sys.getenv}}. \strong{IMPORTANT! Your
password MUST NOT contain the colon character ("\code{:}") or the at-sign
("\code{@}") since these characters have special use as delimiters in the
connection string!}}

\item{protocol}{Default is \code{"sftp://"}. You will probably never change
this value.}

\item{port}{Default is \code{"22"}. You will probably never change this
value. Theoretically, an SFTP server could be set up to respond on a port
other than the default port 22, but this is rarely the case.}
}
\value{
A list of strings which are used by all other \code{sftp_*} functions
when connecting to an SFTP server.
}
\description{
\code{sftp_connect} outputs a list which is used as an argument to all other
functions of the sftp package. Recommended practice is to store the output as
\code{sftp_con} since that is the object name all other sftp functions look
for by default. \strong{IMPORTANT! Your password MUST NOT contain the colon
character ("\code{}") or the at-sign ("\code{@}") since these characters
have special use as delimiters in the connection string!}
}
\examples{

\dontrun{
# Minimal - rely on defaults
sftp_con <- sftp_connect(
  username = Sys.getenv("MY_SFTP_USER"),
  password = Sys.getenv("MY_SFTP_PASS")
)

# Explicit
sftp_con <- sftp_connect(
  server = "mysftp.com",
  folder = "myfolder",
  username = Sys.getenv("MY_SFTP_USER"),
  password = Sys.getenv("MY_SFTP_PASS"),
  protocol = "sftp://",
  port = 22
)
}

}
\author{
\href{https://github.com/stenevang/sftp}{Theodor Stenevang Klemming}
}
