% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp_dir_info.R
\name{sftp_dir_info}
\alias{sftp_dir_info}
\title{Query File Metadata from SFTP Connection}
\usage{
sftp_dir_info(sftp_connection, regexp = NULL)
}
\arguments{
\item{sftp_connection}{A list of connection parameters as created by
\code{\link[coviData:sftp_connect]{sftp_connect()}}}
}
\value{
A \code{tibble} containing the \code{path}, \code{size}, \code{permissions},
\code{modification_time}, and \code{hard_links} of each file in the directory
}
\description{
\code{sftp_dir_info()} returns metadata for all files in the directory specified
by the \code{sftp_connection}.

Note that the year in the \code{modification_time} is
\emph{inferred} such that the resulting datetime is as close as possible to the
current time while remaining in the past or present. This is due to times
returned by \code{\link[curl:curl]{curl()}} not including a year.
}
\keyword{internal}
