% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp.R
\name{sftp_download}
\alias{sftp_download}
\title{Download files from an SFTP account}
\usage{
sftp_download(
  file,
  tofolder = getwd(),
  sftp_connection = get("sftp_con"),
  verbose = TRUE
)
}
\arguments{
\item{file}{A character vector of file names. If the wildcard "*" is used,
then all files on the SFTP url will be downloaded. NOTE! Folders
(directories) will not be downloaded.}

\item{tofolder}{A string containing a valid path - relative or absolute - to
the folder where the files will be saved. Default is the working directory.
If the folder does not exist it will be created.}

\item{sftp_connection}{A list object created with \code{\link{sftp_connect}}.
Default is 'sftp_con'.}

\item{verbose}{Logical. Turn on messages to console. Default is TRUE.}
}
\value{
The function returns the number of files downloaded, following
successful download.
}
\description{
This function is used for downloading files from an SFTP account.
The function uses connection credentials from a list object
created by calling \link{sftp_connect}. The files will be downloaded from
the SFTP account folder where you are currently "standing" as specified in
the connection list object.
}
\examples{

\dontrun{
# minimal - take all files on the SFTP url and save in the working directory
sftp_download("*")

# take a vector of file names from a dataframe
sftp_download(file = files$name, tofolder = "my/relative/path")

# take one specific file name from a vector from a dataframe
sftp_download(file = files$name[1], tofolder = getwd() )
}

}
\seealso{
\link{sftp_upload}, \link{sftp_delete}, \link{sftp_rename}

Other sftp: 
\code{\link{sftp_changedir}()},
\code{\link{sftp_connect}()},
\code{\link{sftp_delete}()},
\code{\link{sftp_listdirs}()},
\code{\link{sftp_listfiles}()},
\code{\link{sftp_list}()},
\code{\link{sftp_makedir}()},
\code{\link{sftp_removedir}()},
\code{\link{sftp_rename}()},
\code{\link{sftp_upload}()}
}
\author{
\href{https://github.com/stenevang/sftp}{Theodor Stenevang Klemming}
}
\concept{sftp}
\keyword{internal}
