% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp.R
\name{sftp_list}
\alias{sftp_list}
\title{List Files and/or Folders in an SFTP Directory}
\usage{
sftp_list(
  sftp_connection = sftp_con,
  verbose = TRUE,
  curlPerformVerbose = FALSE,
  encoding = "UTF-8",
  type = "all",
  recurse = F
)
}
\arguments{
\item{sftp_connection}{A list created by calling \code{\link{sftp_connect}};
default value is \code{sftp_con}.}

\item{verbose}{Logical. Turn on messages to console. Default is \code{TRUE}.}

\item{curlPerformVerbose}{Logical. Turn on messages to console from
\code{\link[RCurl:curlPerform]{RCurl::curlPerform}}.
Default is \code{FALSE}.}

\item{encoding}{Explicitly identifies the encoding of the content that is
returned by the server in its response. Possible values are \code{"UTF-8"}
or \code{"ISO-8859-1"}. Default is \code{"UTF-8"}.}

\item{type}{One of \code{"f"}, code{"file"} for listing files; \code{"d"},
\code{"dir"}, \code{"directory"} for listing directories; or \code{"all"}
for both files and directories, the default.}

\item{recurse}{Logical. Recurse directories? Default is \code{FALSE}.}
}
\value{
A \code{\link[tibble]{tibble}}
}
\description{
\code{sftp_list} lists files, folders, or both in the SFTP location
specified by a list object created using \link{sftp_connect}.
Convenience wrapper functions are available in \code{\link{sftp_listfiles}}
and \code{\link{sftp_listdirs}}
}
\examples{

\dontrun{
# minimal - rely on defaults
files_and_folders <- sftp_list(type = "all")

# explicit
files_and_folders <- sftp_list(sftp_connection = sftp_con,
                               verbose = TRUE,
                               curlPerformVerbose = FALSE,
                               type = "all")
}

}
\seealso{
\code{\link{sftp_listfiles}}, \code{\link{sftp_listdirs}}
}
\author{
\href{https://github.com/stenevang/sftp}{Theodor Stenevang Klemming}
}
