% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp.R
\name{sftp_listfiles}
\alias{sftp_listfiles}
\title{List files in an SFTP directory}
\usage{
sftp_listfiles(
  sftp_connection = sftp_con,
  verbose = TRUE,
  curlPerformVerbose = FALSE,
  recurse = FALSE
)
}
\arguments{
\item{sftp_connection}{A list created by calling \link{sftp_connect}
Default value is "sftp_con".}

\item{verbose}{Logical. Turn on messages to console. Default is TRUE.}

\item{curlPerformVerbose}{Logical. Turn on messages to console from
\code{\link[RCurl]{curlPerform}}. Default is FALSE.}

\item{recurse}{Logical. Recurse directories? Default is FALSE.}
}
\value{
A \code{\link[tibble]{tibble}}
}
\description{
Convenience function wrapping \link{sftp_list} with type hard-set to "file".
}
\examples{

\dontrun{
# minimal - rely on defaults
files <- sftp_listfiles()

# explicit
files <- sftp_listfiles(sftp_connection = sftp_con,
                        verbose = TRUE,
                        curlPerformVerbose = FALSE)
}

}
\seealso{
\link{sftp_listdirs}, \link{sftp_list}

Other sftp: 
\code{\link{sftp_changedir}()},
\code{\link{sftp_connect}()},
\code{\link{sftp_delete}()},
\code{\link{sftp_download}()},
\code{\link{sftp_listdirs}()},
\code{\link{sftp_list}()},
\code{\link{sftp_makedir}()},
\code{\link{sftp_removedir}()},
\code{\link{sftp_rename}()},
\code{\link{sftp_upload}()}
}
\author{
\href{https://github.com/stenevang/sftp}{Theodor Stenevang Klemming}
}
\concept{sftp}
\keyword{internal}
