% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shelby_zip.R
\docType{data}
\name{shelby_zip}
\alias{shelby_zip}
\title{Vector of Shelby County ZIP Codes}
\format{
A \code{character} with 40 values
}
\source{
US Census Bureau 2010 ZCTA to County Relationship File
(\url{https://www.census.gov/geographies/reference-files/2010/geo/relationship-files.html#par_textimage_674173622})
}
\usage{
shelby_zip
}
\description{
Vector of Shelby County ZIP Codes
}
\keyword{datasets}
