% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ael.R, R/utils.R
\name{split_names}
\alias{split_names}
\title{Split Patient Names in AEL Data}
\usage{
split_names(.data)

split_names(.data)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g. a
\code{\link[tibble]{tibble}})}
}
\value{
The original data frame, tibble, etc. with character columns for
\code{PatientLastName}, \code{PatientFirstName}, \code{PatientMiddleName},
\code{PatientMidOthName} inserted after \code{PatientName} (which is a
\code{\link[base]{factor}})

The original data frame, tibble, etc. with character columns for
\code{PatientLastName}, \code{PatientFirstName}, \code{PatientMiddleName},
\code{PatientMidOthName} inserted after \code{PatientName} (which is a
\code{\link[base]{factor}})
}
\description{
\code{split_names} creates columns for the last, first, first-middle, and
other-middle names of patients in AEL data

\code{split_names} creates columns for the last, first, first-middle, and
other-middle names of patients in AEL data
}
