% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize_string}
\alias{standardize_string}
\title{Standardize String Formatting}
\usage{
standardize_string(string, case_fn = stringr::str_to_title, ...)
}
\arguments{
\item{string}{The string or character vector to standardize}

\item{case_fn}{A function for converting to the desired case. Note that
\code{case_fn} is simply the last transformation applied to \code{string}
and may technically perform operations other than case conversion.}

\item{...}{Additional parameters to pass to \code{case_fn}}
}
\value{
A character vector of the same length as \code{string}
}
\description{
\code{standardize_string} ensures that the input string (or character vector)
follows a consistent structure. It coerces the input to type
\code{character}; converts characters to \code{UTF-8} encoding; replaces
non-alphabetical symbols with a space; removes excess whitespace; and
converts to the desired case (title, by default). It also informs the user if
any characters could not be converted to \code{UTF-8}.
}
\details{
This functions is primarily intended to standardize proper nouns (such as
names of persons) into a common format; it doesn't necessarily \emph{correct}
the strings. For more general string standardization of this type, see
the \href{https://tazinho.github.io/snakecase/}{snakecase} package.
}
