% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_names.R
\name{std_names}
\alias{std_names}
\title{Parse Investigator Names and Remove (Most) Non-Alphabetical Characters}
\usage{
std_names(string, case = c("upper", "lower", "title", "sentence"))
}
\arguments{
\item{string}{A character vector containing investigator names}

\item{case}{Case of output}
}
\value{
The input \code{string} with cleaned names
}
\description{
\code{std_names()} is designed to parse names of individuals into a standard
format. It:
\enumerate{
\item Replaces square brackets and curly braces with parentheses
\item Removes parenthetic expressions
\item Removes ordinal and numeric expressions
\item Removes symbols (replaces all except apostrophes with a space)
\item Removes extraneous whitespace
\item Converts to the desired case
}
}
