% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_names.R
\name{str_remove_symbols}
\alias{str_remove_symbols}
\title{Remove Symbols from String}
\usage{
str_remove_symbols(string)
}
\arguments{
\item{string}{A character vector}
}
\value{
The input \code{string}, with symbols removed
}
\description{
\code{str_remove_symbols()} removes all non-alphanumeric characters from a string.
It removes apostrophes entirely but replaces other symbols with a space.
}
\keyword{internal}
