% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_names.R
\name{str_to_ascii}
\alias{str_to_ascii}
\title{Convert a String to ASCII Representation}
\usage{
str_to_ascii(string, trans = TRUE)
}
\arguments{
\item{string}{A character vector}

\item{trans}{Should non-ASCII characters be transliterated?}
}
\value{
An ASCII-encoded character vector
}
\description{
\code{str_to_ascii()} converts a string to ASCII characters only. When
\code{trans = TRUE}, non-ASCII characters are transliterated; otherwise, non-ASCII
characters are replaced with the ASCII SUBSTITUTE CHARACTER (\code{0x1A}).
}
