% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{str_to_factor}
\alias{str_to_factor}
\title{Convert Character Strings to Factors}
\usage{
str_to_factor(string, encoding = FALSE)
}
\arguments{
\item{string}{A character vector to convert}

\item{encoding}{A logical indicating whether to convert the character
encoding, or a string indicating the encoding to convert to. If
\code{encoding == TRUE}, then characters are converted to 'UTF-8' encoding.}
}
\value{
A factor
}
\description{
\code{str_to_factor} converts a character (string) vector to a factor after
removing extra whitespace, converting to title case, and (optionally)
converting the encoding a specified standard.
}
