% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{str_to_vb_html}
\alias{str_to_vb_html}
\title{Convert String to HTML Suitable for Inclusion in VB Script}
\usage{
str_to_vb_html(string, collapse = NULL)
}
\arguments{
\item{string}{A character vector}

\item{collapse}{An optional character string to separate the results. Not
\code{NA_character_}.}
}
\value{
An HTML string
}
\description{
Convert String to HTML Suitable for Inclusion in VB Script
}
