% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_report_date.R
\name{to_report_date_fst}
\alias{to_report_date_fst}
\title{Convert NBS Data to \code{fst} Files Containing Report Dates}
\usage{
to_report_date_fst(path, birth_date, save_as)
}
\arguments{
\item{path}{The paths to data files to read}

\item{birth_date}{The date of creation of each file - used as inferred report
date}

\item{save_as}{The location to save each file}
}
\value{
A \code{tibble} containing the input vectors
\code{path}, \code{birth_date}, \code{save_as} (invisibly)
}
\description{
\code{to_report_date_fst()} is an internal function called by
\code{\link[coviData:archive_report_date]{archive_report_date()}}. It saves
inferred report dates (\code{birth_date}) to \code{save_as} for each NBS ID in the
files in \code{path}.
}
\keyword{internal}
