% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim_backups}
\alias{trim_backups}
\title{Remove Files from Directory Based on Date Created}
\usage{
trim_backups(directory, pattern = NULL, n_recent = 7)
}
\arguments{
\item{directory}{The path to the backup directory}

\item{pattern}{The naming pattern of the backup files, as a regular
expression. This allows maintenance of backups when other files are also
present in the \code{directory} (though this is not recommended).}

\item{n_recent}{The number of most recent files to keep}
}
\value{
The deleted paths (invisibly)
}
\description{
\code{trim_backups()} deletes old backup files from a directory. More
specifically, it identifies files in a \code{directory} with a common naming
\code{pattern} and it keeps the most recent \code{n_recent} of those files
(deleting the rest).
}
