% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vac_prep.R
\name{vac_distinct}
\alias{vac_distinct}
\title{Deduplicate Vaccination Data}
\usage{
vac_distinct(data, cols = "asiis_pat_id_ptr", desc = TRUE)
}
\arguments{
\item{data}{Vaccination data}

\item{cols}{\code{\link[dplyr:dplyr_data_masking]{<data masking>}}
\code{character} vector of variables to use when determining uniqueness. If
there are multiple rows for a given combination of inputs, only the first
row will be preserved. The default de-duplicates by person.}

\item{desc}{Should \code{dose_count} be
\code{\link[dplyr:arrange]{arrange}}\code{d} in descending order?}
}
\value{
An object of the same type as \code{data}. The output has the following
properties: \itemize{
\item Rows are a subset of the input but appear in the same order
\item Columns are not modified
\item Groups are not modified
\item Data frame attributes are preserved
}
}
\description{
Create distinct vaccination data. Sorts by \code{dose_count} prior to calling
\code{\link[dplyr:distinct]{distinct()}}.
}
\keyword{internal}
