% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vac_prep.R
\name{vac_filter_doses}
\alias{vac_filter_doses}
\title{Filter to Valid Doses in Vaccination Data}
\usage{
vac_filter_doses(data)
}
\arguments{
\item{data}{Vaccination data with \code{numeric} \code{dose_count} and \code{max_doses}
columns}
}
\value{
\code{data} filtered to valid doses
}
\description{
Filter to Valid Doses in Vaccination Data
}
\keyword{internal}
