% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vac_prep.R
\name{vac_filter_residents}
\alias{vac_filter_residents}
\title{Filter to Possible Shelby Residents in Vaccination Data}
\usage{
vac_filter_residents(data)
}
\arguments{
\item{data}{Vaccination data with \code{logical} \code{resident} column}
}
\value{
\code{data} filtered to potential residents
}
\description{
Filter to Possible Shelby Residents in Vaccination Data
}
\keyword{internal}
