% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{vac_load}
\alias{vac_load}
\title{Deprecated - Use \code{read_vac()} (now the backend of this function)}
\usage{
vac_load(date = NULL, ext = c("csv", "xlsx"), path = character(), ...)
}
\arguments{
\item{date}{The download date of the data}

\item{ext}{Deprecated}

\item{path}{Deprecated}

\item{...}{Additional arguments to pass to
\code{\link[coviData:read_vac]{read_vac()}}}
}
\description{
Deprecated - Use \code{read_vac()} (now the backend of this function)
}
\keyword{internal}
