% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vac_prep.R
\name{vac_mutate}
\alias{vac_mutate}
\title{Create and Transform Variables in Vaccination Data}
\usage{
vac_mutate(data)
}
\arguments{
\item{data}{Vaccination data from TennIIS}
}
\value{
The data with parsed \code{address_zip}, \code{integer} \code{dose_count}, and
added columns \code{resident} (\code{logical}), \code{max_doses} (\code{integer}),
\code{recip_fully_vacc} (\code{logical})
}
\description{
Create and Transform Variables in Vaccination Data
}
\keyword{internal}
