% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vac_parse.R
\name{vac_parse_zip}
\alias{vac_parse_zip}
\title{Parse Zip Codes from Vaccination Data}
\usage{
vac_parse_zip(string)
}
\arguments{
\item{string}{A \code{character} vector of ZIP codes}
}
\value{
A \code{character} vector of 5-digit ZIP codes. Invalid values are
converted to missing; valid ZIP codes outside Shelby County are converted
to \code{"Other"}.
}
\description{
\code{vac_parse_zip()} parses ZIP codes from the \code{address_zip} field in
vaccination data. It extracts the first 5-digit sequence in the field and
compares with a list of valid Shelby County ZIP codes.
}
