% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_load.R
\name{var_load}
\alias{var_load}
\title{Load Variant Cases from NIT Project}
\usage{
var_load(api_token = Sys.getenv("redcap_NIT_token"))
}
\arguments{
\item{api_token}{Character. An API token/key for the NIT project.}
}
\value{
A \code{tibble} of \code{character} columns containing all variant case records
}
\description{
\code{var_load()} loads variant cases from the NIT REDcap project. If possible,
it downloads the most up-to-date information from the REDcap server; if not,
it reads from the most recently downloaded NIT file.
}
