% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_parse.R
\name{var_parse}
\alias{var_parse}
\title{Parse Variant Data from NIT Project}
\usage{
var_parse(data = var_load())
}
\arguments{
\item{data}{NIT variant data}
}
\value{
A \code{tibble} of parsed data
}
\description{
\code{var_parse()} deduplicates variant data from the NIT REDcap project and
parses key variables into a standard format. Currently parsed variables are:
\itemize{
\item{\code{nbs}}
\item{\code{firstname}}
\item{\code{lastname}}
\item{\code{dob}}
\item{\code{specimendate}}
}
}
