% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{vb_newline_notify}
\alias{vb_newline_notify}
\title{Replace LF and CRLF with VB Constants for \code{notify()}}
\usage{
vb_newline_notify(string)
}
\arguments{
\item{string}{A string to perform replacement on}
}
\value{
The string with (CR)LF replaced with "& VbCrLf &"
}
\description{
Replace LF and CRLF with VB Constants for \code{notify()}
}
