% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{vec_sample}
\alias{vec_sample}
\title{Re-implementation of \code{base::sample()} Using the vctrs Package}
\usage{
vec_sample(x, n, replace = FALSE, weight_by = NULL)
}
\arguments{
\item{x}{The data to sample}

\item{n}{The number of samples to return}

\item{replace}{Should sampling be performed with (\code{TRUE}) or without
(\code{FALSE}, the default) replacement?}

\item{weight_by}{Sampling weights. This must evaluate to a vector of
non-negative numbers the same length as the input. Weights are
automatically standardised to sum to 1.}
}
\description{
\code{\link[base:sample]{sample()}} behaves unexpectedly when called on
objects other than atomic vectors. This re-implementation has identical logic
but uses the \href{https://vctrs.r-lib.org}{vctrs} package to generalize to
other inputs. Parameters are modeled after
\code{\link[dplyr:slice_sample]{slice_sample()}}.
}
