% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sas.R
\name{write_sas_nbs}
\alias{write_sas_nbs}
\title{Separate +/- Investigation/PCR Data Files for SAS}
\usage{
write_sas_nbs(
  pcr = process_pcr(inv = inv),
  inv = process_inv(),
  dir = "V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/Data for SAS",
  date = attr(pcr, "date"),
  force = FALSE
)
}
\description{
Separate +/- Investigation/PCR Data Files for SAS
}
