% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sas.R
\name{write_sas_neg}
\alias{write_sas_neg}
\title{Write Negative Investigation and PCR Data to CSV for SAS}
\usage{
write_sas_neg(
  data = neg(process_pcr()),
  type = c("inv", "pcr"),
  dir = "V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/Data for SAS",
  date = attr(data, "date"),
  force = FALSE
)
}
\description{
Write Negative Investigation and PCR Data to CSV for SAS
}
