% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sas.R
\name{write_sas_pos}
\alias{write_sas_pos}
\title{Write Positive Investigation and PCR Data to XLSX for SAS}
\usage{
write_sas_pos(
  data = pos(process_inv()),
  type = c("inv", "pcr"),
  dir = "V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/Data for SAS",
  date = NULL,
  force = FALSE
)
}
\description{
Write Positive Investigation and PCR Data to XLSX for SAS
}
