% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_date.R
\name{add_collection_date}
\alias{add_collection_date}
\title{Add Specimen Collection Date to a Report Date Tibble}
\usage{
add_collection_date(
  .data,
  .from_col = "specimen_coll_dt",
  .id_col = "inv_local_id",
  ...,
  date = Sys.Date(),
  from_file = path_create("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP Data/",
    "Sandbox data pull Final/")
)
}
\arguments{
\item{.data}{A \code{tibble} returned from \code{coalesce_report_date()}}

\item{.from_col}{A string indicating the variable in \code{.from_file} to use as
specimen collection date}

\item{.id_col}{A string indicating the variable in \code{.data} and \code{.from_file}
to use as an ID variable}

\item{...}{\verb{<tidy-select>} Additional variables to add}

\item{.date}{The date of the NBS file to use for collection dates}

\item{.from_file}{A file path to the NBS data with the desired specimen
collection dates; if only a directory is provided, the NBS file
corresponding to \code{date} will be used}
}
\description{
\code{add_collection_date()} reads the specimen collection date column from an
NBS data file and adds it to the input data (typically the result of
\code{coalesce_report_date()}).
}
