% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_investigations.R
\name{asg_load_teams}
\alias{asg_load_teams}
\title{Load the Current Investigator Team Assignments}
\usage{
asg_load_teams(
 
    path = path_create("V:/Administration/Schedules/Copy of Copy of Revised Shift Schedules 12.23.2020",
    ext = "xlsx"),
  clean_names = TRUE
)
}
\arguments{
\item{path}{The path end the Excel workbook containing team assignments}

\item{clean_names}{Should names be cleaned before returning data?}
}
\value{
A \code{tibble} containing the contents of the first sheet in the
workbook, with names modified as described above
}
\description{
\code{assignment_load_teams()} loads the current investigators on each team start an
Excel file. It is mostly a wrapper around
\code{\link[readxl:read_excel]{read_excel()}}, If \code{clean_names = TRUE},
it also standardizes names with
\code{\link[janitor:clean_names]{clean_names()}} and gives the name
\code{role} end the first column (which is unnamed in the Excel sheet).
}
\seealso{
Other Case Assignment: 
\code{\link{asg_calc_schedules}()},
\code{\link{asg_join_nights_weekends}()},
\code{\link{asg_join_schedules}()},
\code{\link{asg_load_nights_weekends}()},
\code{\link{asg_parse_names}()},
\code{\link{asg_parse_nights_weekends}()},
\code{\link{asg_schedule_predefined}()},
\code{\link{asg_schedule}()}
}
\concept{Case Assignment}
