% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coalesce_by_column}
\alias{coalesce_by_column}
\title{Coalesce Rows by Columns in Data Frame}
\usage{
coalesce_by_column(x)
}
\arguments{
\item{x}{A grouped data frame}
}
\description{
\code{coalesce_by_column()} is the workhorse behind \code{coalesce_dupes()}; when
combined with \code{\link[dplyr:group_by]{group_by()}} and
\code{\link[dplyr:summarise]{summarize()}}, it allows coalescing of data
frame \emph{rows}.
}
\references{
See Jon Harmon's solution to
\href{https://stackoverflow.com/questions/45515218/combine-rows-in-data-frame-containing-na-to-make-complete-row}{this question}
on Stack Overflow.
}
