% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_generic.R
\name{download_interview_report}
\alias{download_interview_report}
\title{Download a Report from the Integrated Data Tool REDcap Project}
\usage{
download_interview_report(
  date = Sys.Date(),
  api_token,
  report_id,
  directory,
  new_file,
  force = FALSE
)
}
\arguments{
\item{date}{A \code{Date} indicating the date of the file to download}

\item{api_token}{The API token for accessing the Data for Regions REDcap
project. This should be stored in an \emph{.Renviron} file; see
\link{using-renviron} for details.}

\item{report_id}{The ID of the report to download}

\item{directory}{A string specifying the save directory; this should usually
be left alone}

\item{new_file}{A string specifying the save file name; this should always
end in \emph{.csv} and should usually be left alone}

\item{force}{A logical indicating whether to ignore any existing files
matching \code{date} and \code{directory}}
}
\value{
Invisibly returns the path to the new data file
}
\description{
\code{download_interview_report()} downloads a report from the Integrated Data
Tool project. You'll need API access to the project (and an API token for it)
to use this function.
}
\keyword{internal}
