% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_generic.R
\name{download_redcap_file}
\alias{download_redcap_file}
\title{Download REDcap File from TN REDcap}
\usage{
download_redcap_file(
  date = Sys.Date(),
  api_token,
  redcap_file,
  directory,
  new_file,
  convert = FALSE,
  force = FALSE
)
}
\arguments{
\item{date}{A \code{Date} indicating the date of the file to download}

\item{api_token}{The API token for the REDcap project to access. Ideally,
this should be stored in Renviron and not in a script.}

\item{redcap_file}{A string indicating the name of the file to download from
REDcap}

\item{directory}{A string indicating the directory to save the file in}

\item{new_file}{A string indicating the name to save the file under}

\item{convert}{A logical indicating whether the user would like to convert
the file to another format. If so, the file will open in Excel after
downloading.}

\item{force}{A logical indicating whether to ignore existing files with
the given date already in the directory}
}
\description{
Download REDcap File from TN REDcap
}
