% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{dttm_to_dt}
\alias{dttm_to_dt}
\title{Assign Date Type to DateTime Variables}
\usage{
dttm_to_dt(.x)
}
\description{
\code{dttm_to_dt} is an opinionated formatter for dates and datetimes. It
prefers simple dates to datetimes, and checks any datetime variables for
additional information in the hour:minute:second portion of the variable. If
it finds none, it converts the variable to a standard date.
}
