% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{ennotify}
\alias{ennotify}
\title{\code{entrace()} + \code{notify()}}
\usage{
ennotify()
}
\description{
\code{ennotify()} adds a backtrace to base R errors and sends a notification
email with the failing function an a full backtrace. However, it cannot
capture the condition object associated with the error, and thus cannot
show error messages. Logging errors is recommended.
}
\details{
Set \code{ennotify()} as your error handler using

\verb{options(error = quote(coviData::ennotify()),}
\verb{ rlang_backtrace_on_error = "full")}

The function currently emails me (Jesse Smith) every time an error is
processed, so please do not use this before consulting me first.
}
