% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_file}
\alias{find_file}
\title{Find a File with the Specified Date + Pattern in the File Name}
\usage{
find_file(
  date = Sys.Date(),
  pattern = paste0(".*", date, ".*"),
  directory = NULL,
  file_name = NULL,
  date_flag = NULL,
  rtn_error = TRUE
)
}
\arguments{
\item{date}{A \code{Date} indicating the date in the filename}

\item{pattern}{The pattern to match when searching for the filename. The
default is to look for any file with \code{date} in the name.}

\item{directory}{The directory in which to search for the file.}

\item{file_name}{If the search does not return a file known to exist,
\code{filename} can be used to specify the file directly}

\item{date_flag}{A string used to print more informative messages in some
functions}
}
\value{
A string containing the full path to the file that was found
}
\description{
\code{find_file} looks for a file in the folder \code{directory}
with the date \code{date} in its name. It returns the path to that file; if
more than one file is found, it returns the path to the first one and issues
a warning.
}
\keyword{internal}
