% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ael.R
\name{load_ael}
\alias{load_ael}
\title{Load, Clean, & Merge AEL Data Files for 2 Consecutive Days}
\usage{
load_ael(
  day1 = Sys.Date(),
  dir_path = "V:/EPI DATA ANALYTICS TEAM/AEL Data/",
  day2 = day1 - 1,
  day1_file = NULL,
  day2_file = NULL,
  overwrite_names_1 = FALSE,
  overwrite_names_2 = FALSE,
  string_to_factor = TRUE,
  encoding = TRUE,
  check = TRUE
)
}
\description{
\code{get_ael} prepares AEL data for analysis. It imports the data,
cleans name variables, restricts to our jurisdiction, and merges the files
for use with NBS.
}
