% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_generic.R
\name{load_data}
\alias{load_data}
\title{Load a Data File Using the Directory, Date, and Extension}
\usage{
load_data(
  date,
  directory = "V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/MSR PCR/",
  file_name = NULL,
  ext = c("csv", "xlsx"),
  pattern = paste0(".*", date, ".*", ext[[1]])
)
}
\description{
\code{load_data} is meant to find and load a delimited or Excel data file
given the directory that it's in, the date of interest, and the file
extension.
}
