% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{notify}
\alias{notify}
\title{Send an Email Notification with Outlook}
\usage{
notify(to, subject, body = NULL, cc = NULL, html = FALSE)
}
\arguments{
\item{to}{A character vector containing recipient email addresses. Required.}

\item{subject}{A character string containing the email subject. Required.}

\item{body}{A character string containing the email body; if \code{html = TRUE},
this will be parsed as HTML. Optional.}

\item{cc}{A character vector containing email address to "cc". Optional.}

\item{html}{Should the \code{body} be treated as plain text or HTML? The default
is plain text (\code{FALSE}).}
}
\description{
\code{notify} sends a notification email via Outlook. This assumes that Outlook
is installed on your machine and that you've already set up an account.
}
\keyword{internal}
