% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{notify}
\alias{notify}
\title{Send an Email Notification with blastula}
\usage{
notify(to, subject = "", body = "", html = FALSE, fail = FALSE)
}
\arguments{
\item{to}{Character. The recipient(s) of the email.}

\item{subject}{Character. The subject of the email.}

\item{body}{Character. The body of the email. If \code{html = FALSE}, this should
be plain text. If \code{html = TRUE}, this should be HTML text.}

\item{html}{Is the \code{body} HTML or plain text? All messages are sent in HTML;
this argument determines how messages are parsed.}

\item{fail}{Should failure cause an error?}
}
\description{
\code{notify} sends a notification email using a Visual Basic interface to Outlook
}
\keyword{internal}
