% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_generic.R
\name{preprocess}
\alias{preprocess}
\title{Standardize Variables, Handle Dates, and Remove Bad Columns}
\usage{
preprocess(data, na_constant = TRUE)
}
\arguments{
\item{data}{A dataframe(-like) object}

\item{na_constant}{Should \code{NA} values be considered when deciding
whether a column is constant? The default is \code{TRUE}, which will
keep columns that only have 1 unique non-missing value.}
}
\value{
A processed tibble
}
\description{
\code{preprocess} standardizes column names, converts variables to the
appropriate data types (including oddly formatted dates/datetimes), and
removes empty or constant columns.
}
