% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ael.R
\name{read_ael}
\alias{read_ael}
\title{Read AEL Data from Excel File}
\usage{
read_ael(path, string_to_factor = FALSE, encoding = TRUE)
}
\arguments{
\item{path}{The file path to the Excel file}

\item{string_to_factor}{A logical indicating whether to encode character
columns as factors or leave as character}

\item{encoding}{A logical indicating whether to convert the character
encoding, or a string indicating the encoding to convert to. If
\code{encoding == TRUE}, then characters are converted to 'UTF-8' encoding.}
}
\value{
A \code{\link[tibble]{tibble}} containing the contents of the Excel
file
}
\description{
\code{read_xl} reads in an xls(x) file, converts date columns to
\code{Date} class, and optionally converts \code{character} columns to
\code{factor} columns. \code{read_ael} uses all rows of a file to guess the
column type (technically, the first 21,478,836 rows on 64-bit R). It is
intended for reading an Excel file containing data sent from AEL to the
Shelby County Health Department.
}
