% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{read_file}
\alias{read_file}
\title{Read a Tabular Data File}
\usage{
read_file(
  path,
  file_type = c("auto", "delimited", "excel"),
  msg = "Reading file..."
)
}
\description{
Read a Tabular Data File
}
