% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_file}
\alias{read_file}
\title{Read a Tabular Data File}
\usage{
read_file(
  path,
  file_type = c("auto", "delimited", "excel"),
  type_convert = TRUE,
  msg = "Reading file..."
)
}
\arguments{
\item{path}{A string indicating the path to the data file; will be expanded
if necessary}

\item{file_type}{A string indicating the type of file to read; only
"delimited" and "excel" files are supported. The default is "auto", which
determines the file type from the file extension.}

\item{type_convert}{Should \code{read_file()} attempt to guess the data type of
the columns after reading?

\emph{Note: The default is \code{TRUE} for compatibility; want to transition to
\code{vroom::vroom()} as backend and swap default to \code{FALSE}.}}

\item{msg}{A message to be displayed prior to beginning a file read; for use
inside other functions}
}
\value{
A tibble containing the read data
}
\description{
\code{read_file} reads delimited- and Excel-type data files. The backend for
delimited files is \code{\link[data.table]{fread}}; the backend for Excel
files is \code{\link[readxl]{read_excel}}.
}
\keyword{internal}
