% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace_ael}
\alias{replace_ael}
\title{Add Updated Names Columns to AEL File and Replace Existing File}
\usage{
replace_ael(
  date = Sys.Date(),
  directory = "V:/EPI DATA ANALYTICS TEAM/AEL Data/",
  date_file = NULL,
  overwrite_names = FALSE
)
}
\description{
Add Updated Names Columns to AEL File and Replace Existing File
}
