% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ael.R
\name{replace_ael}
\alias{replace_ael}
\title{Add Updated Names Columns to AEL File and Replace Existing File}
\usage{
replace_ael(
  day = Sys.Date(),
  dir_path = "V:/EPI DATA ANALYTICS TEAM/AEL Data/",
  day_file = NULL,
  overwrite_names = FALSE,
  encoding = TRUE
)
}
\description{
Add Updated Names Columns to AEL File and Replace Existing File
}
