% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_names.R
\name{std_names}
\alias{std_names}
\title{Parse Investigator Names and Remove (Most) Non-Alphabetical Characters}
\usage{
std_names(string)
}
\arguments{
\item{string}{A character vector containing investigator names}
}
\value{
The input \code{string} with cleaned names
}
\description{
\code{std_names()} is designed to parse names of individuals into a standard
format. It:
\enumerate{
\item Replaces square brackets and curly braces with parentheses
\item Removes parenthetic expressions
\item Removes ordinal and numeric expressions
\item Converts \code{NA}s to text (for the next function)
\item Passes the intermediate to
\code{\link[janitor:make_clean_names]{
      make_clean_names(
        case = "title",
        use_make_names = FALSE,
        transliterations = c("Any-Latin", "Latin-ASCII"),
        sep_out = " "
      )
    }}
\item Removes leading and trailing dashes
\item Removes extraneous whitespace
\item Inverts converted \code{NA}s (back to \code{NA_character_})
}
}
