% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vac_load.R
\name{vac_load}
\alias{vac_load}
\title{Load Vaccination Data}
\usage{
vac_load(date = NULL, ext = c("csv", "xlsx"), path = character(), ...)
}
\arguments{
\item{date}{A \code{Date} or string in the format "YYYY-MM-DD"}

\item{ext}{The file type to return. Files may be saved as "fst", "csv", or
"xlsx".}

\item{path}{Optional \code{character}. The path to the file to load; if this is
supplied, \code{date} and \code{ext} are ignored.}

\item{...}{Additional arguments to pass to \code{\link[vroom:vroom]{vroom()}}
(if loading a csv file) or \code{\link[readxl:read_excel]{read_excel()}}
(if loading an xlsx file)}
}
\value{
A \code{tibble} containing the loaded data. Unless otherwise specified in
\code{...}, all variables are read as character (which means that Excel dates
will be in Excel's numeric representation).
}
\description{
\code{vac_load()} loads vaccination data from a CSV or Excel file. By default,
it loads the most recent data available; this can be controlled by supplying
the \code{date} of the data to load.
}
