% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vac_parse.R
\name{vac_prep}
\alias{vac_prep}
\title{Prepare Vaccine Data for Analysis}
\usage{
vac_prep(.data, distinct = FALSE)
}
\arguments{
\item{.data}{Vaccination data, as loaded by
\code{\link[coviData:vac_load]{vac_load()}}}

\item{distinct}{Should data be de-duplicated to distinct individuals?}
}
\value{
A \code{tibble}
}
\description{
\code{vac_prep()} prepares vaccination data for analysis. It performs the
following steps:
\itemize{
\item{Standardize variable names}
\item{Extract ZIP codes; convert non-Shelby to \code{"Other"}, invalid to \code{NA}}
\item{Determine Shelby County residency based on ZIP}
\item{Convert \code{dose_count} to \code{integer}}
}
}
